chore: refactor components folders
