--- packages/storybook/src/templates/form-wmebv/components/Logo.tsx
@@ -1,572 +1,0 @@
-import { Link } from '@utrecht/component-library-react';
-import { useId } from 'react';
-
-export const PageHeaderLogo = () => {
-  const homepageTitleId = useId();
-  return (
-    <Link boxContent href="/" aria-labelledby={homepageTitleId}>
-      <span id={homepageTitleId} hidden>
-        homepage
-      </span>
-      <svg
-        role="img"
-        aria-label="logo van gemeente Voorbeeld"
-        width="177"
-        height="64"
-        viewBox="0 0 177 64"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M32.8349 3.00852L33.5091 2.69139L33.8817 2.60354H35.3344L36.7346 2.28731L36.1671 1.987L32.764 2.10937L32.5526 2.02263L32.3213 2.21694L32.231 2.56746L32.4794 2.9182L32.8349 3.00852ZM26.2037 4.81847L28.1533 3.95742L28.6501 3.83505L29.2879 3.55334L29.87 3.53586L30.0664 3.43254L30.0312 3.3429L29.9955 3.16585L29.9612 3.07799L30.7935 3.04348L31.1835 2.90341L31.5205 2.44352L29.43 2.84827L29.2333 2.99149L29.002 3.18288L27.1062 3.5697H26.7885L26.5387 3.80009L25.7762 4.06342L25.6181 4.46683L25.7938 4.81824L26.2037 4.81847ZM39.1643 2.23555V2.30502L39.2533 2.4442L39.537 2.09233L41.9662 1.33573H42.1611V1.17884L41.7695 1.05715L40.3711 1.17884L39.9984 1.26669L39.4668 1.49619L39.1113 1.81152L39.2533 2.12909L39.1643 2.23555ZM45.7237 0.474226L45.7765 0.317347L45.884 0.281487L45.9177 0.230838H45.599L45.4561 0.352532L45.4224 0.474452L45.6345 0.581578L45.7237 0.474226ZM43.8808 0.335498L44.3792 0.507845L44.4131 0.419541L44.4501 0.386151L44.5021 0.350962H44.5734L43.7915 0L43.6131 0.0192735L43.6312 0.174585L43.8808 0.335498ZM22.3573 7.77611L22.676 7.31645L23.7937 6.24272L24.2002 6.08584L24.3768 5.98117H24.7486V5.8243L24.0426 5.71941L23.0834 6.2797L22.1617 7.12371L21.5587 7.89937L22.0011 8.09301L22.3573 7.77611ZM54.1436 9.13044L54.2685 8.74249L54.303 8.37562L54.2856 8.00583L54.2493 7.35253H54.0925L53.0613 7.0883L52.6356 6.84334L52.6024 6.45495L52.5322 6.26042V6.15643L52.5848 6.08584L52.6887 5.98117V5.8243H52.6015L52.4766 5.94599L52.3169 5.99866L52.1403 5.96504L52.0177 5.8243L51.7692 5.89377L51.2008 5.5762L50.7561 5.43523L50.6526 5.33259L50.5815 5.22568L50.5083 5.17234L50.4914 5.1199L50.1526 4.87337L50.0824 4.83976L49.6031 2.90453L49.4312 2.68669L49.1489 2.54415L47.7468 2.3169L44.4147 2.82609L42.3068 3.4424L41.9534 3.72165L41.8296 3.75639L41.6857 3.70349L41.4386 3.49305L41.2957 3.44083L38.5302 3.53048L37.1656 3.59973L36.2084 4.09166L35.8895 4.10959L35.3748 4.47916L33.1233 5.37763L32.6249 5.79964L31.9516 5.90475L31.6684 6.00964L31.42 6.27432L31.2061 6.59144L30.9753 6.85656L30.4626 7.08382L30.1602 7.34805L29.6633 7.9294L29.381 8.63088L28.9891 10.2156L28.7057 10.7956L28.4234 10.991L27.7666 11.1479L27.4997 11.3604L27.2885 11.6044L24.4482 14.1192L23.898 14.3658L23.3305 14.437L23.029 14.5437L22.428 15.0527L22.1082 15.2288L21.5571 15.2624L20.9733 15.0879L20.6526 14.6658L20.8647 13.9615L20.7592 13.7871L20.6187 13.9263L19.7487 13.7871L19.5906 13.9789L19.4849 14.4534L19.4321 14.9982L19.4147 15.4209L19.288 16.1773L18.687 17.7252L18.5438 18.4467L18.5095 18.7974L18.2611 19.9426L17.6571 24.3902L17.373 24.9888V25.1632L17.2481 25.8151L16.3964 28.3153L14.9603 30.883L13.24 33.3989L11.2195 35.8254L10.7421 36.1441L10.1919 36.2313L9.53695 36.0921L9.4466 36.2474L9.35627 36.4245L9.51662 36.8122L9.78177 37.7102L9.8371 38.0616L9.80096 38.1313L9.72892 38.221L9.65868 38.3433L9.62322 38.5195L9.65868 38.6605L9.72892 38.7501L9.81813 38.8212L10.1716 39.3129L10.28 39.5253L10.1013 39.6645L9.96019 39.7349L9.85449 39.7539L9.79894 39.7349L9.40912 39.5253L8.93077 39.4177L8.66404 39.4377L8.54095 39.5946L7.65562 39.8595L7.24909 40.139L7.37286 40.7564H7.51605L7.62242 40.7385L7.67527 40.6488L7.95848 40.5784H8.31284L8.6681 40.4744L8.93325 40.2799L9.34227 40.5601L9.67833 40.8774L9.90779 41.3528L9.99633 42.0894L10.191 42.3888L11.1658 42.6889L11.5195 42.9355L11.6615 43.0941L11.1823 43.128L10.9338 43.232L10.7563 43.4274L10.7224 43.6199V43.8664L10.7563 44.13L10.8272 44.3245L10.9338 44.4467L11.075 44.5531L11.2179 44.6056L11.8375 44.6246L12.0143 44.7974L12.0852 45.2212L11.7464 44.9417L11.3749 44.835L10.1169 44.7638L9.74428 44.8687L8.99897 45.2212L10.1167 46.2571L10.4521 46.4532L10.666 46.4868L11.1089 46.4341L11.3218 46.4532L11.9068 46.7151L12.5462 46.8212L12.936 47.048L13.2375 47.4178L13.3604 47.9121L13.1673 48.422L12.742 48.5437L12.2819 48.5269L11.9978 48.6134L11.7128 48.474L10.9858 48.5789L10.6513 48.474L10.4374 48.193L10.1531 47.4178L9.96109 47.0641L9.37547 46.5536L8.68504 46.2732L7.33785 46.1331V45.9922L7.42616 45.9025L7.67459 45.7456L7.81778 45.6936L7.69289 45.359L7.49798 45.0789L7.26761 44.8673L7.05531 44.7425L6.73641 44.7089L2.94214 45.1123L1.48925 46.0108L1.34629 46.2042L1.39914 46.5737L1.55926 46.8203L2.02113 47.225L2.19706 47.4866L2.39197 47.8389L2.60495 48.156L2.85338 48.2992L3.65131 48.4731L3.97088 48.4554L4.55583 48.2432L4.76813 48.2609L5.3183 48.7907L6.16998 49.212L6.5598 49.6154L6.70118 49.6851L6.85927 49.6672L7.03996 49.5442L7.55263 49.4917L7.76584 49.3861L7.92528 49.1598L8.35056 48.3693L8.54366 48.2436L9.13087 48.2956L9.36237 48.4023L10.1429 49.0706L10.427 49.2461L10.7288 49.3335L12.022 49.3525L12.34 49.2629L12.5003 49.0513L12.6406 49.1564L12.676 49.3308V49.5614L12.7115 49.8409L12.8183 50.1053L12.8727 50.1748L13.2793 50.1941L13.4387 50.2826L13.7931 50.4742L13.9512 50.5269L14.0569 50.5623L14.608 50.5975H14.7136L14.9804 50.4572L14.9987 50.3675L15.0515 50.0672L15.0332 50.0338L14.9614 49.8568L14.6436 49.489L14.608 49.4366L14.5379 49.2956V49.0668L14.6283 48.8904L14.6619 48.7335L14.486 48.5392L15.2313 48.1172L15.7094 47.9426L15.9764 47.8359L16.3472 47.8026H16.3998L16.5767 47.8359L16.6824 48.0329L16.665 48.3124V48.4007L16.5419 48.6811L16.5229 48.786L16.5055 48.8906L16.5932 48.9603L16.7707 49.0845L16.8581 49.1031H16.9484L17.055 49.0845L17.3574 49.0307L17.782 49.1387H17.8349L18.0307 49.1197L18.2256 49.0484L18.3686 48.9435L19.4673 47.6082L19.929 47.3617L20.1221 47.3803L20.2305 47.4139L20.3716 47.4666L20.4419 47.5009L20.6176 47.6266L20.7608 47.818L20.7971 48.1542L20.7413 48.3664L20.7079 48.4886L20.6704 48.7178L20.653 48.8227L20.7079 49.0323L20.7608 49.2095L20.1404 48.9276L19.9823 48.9467L19.8931 49.2125L20.1943 49.3335L21.1881 49.3687L21.3136 49.4023L21.4548 49.4384L21.7199 49.5623L22.0034 49.6842L22.3049 49.5966L22.465 49.4191L22.6602 49.227L22.8743 48.9276L22.9797 48.7875L23.0674 48.6439L23.1941 48.4343L23.2469 48.2934L23.2995 48.1685L23.3361 48.0282L23.3551 47.957L23.4425 47.833L23.6559 47.9758L23.6749 48.011L23.6986 48.0253L24.123 48.3046L24.2287 48.4984L24.4062 48.8346L24.3534 49.2044L24.1953 49.5551L24.177 49.7297L24.1424 50.0276L24.2298 50.2402L24.7079 50.7665L24.78 50.9612L24.9202 51.4543L24.9751 51.576V51.5952H24.9923L25.2247 51.7187L25.7721 51.6649H25.8286L26.0215 51.6827L26.2713 52.0011V52.7907L26.5197 52.8624L27.0509 52.7387L28.7179 52.81L29.0183 52.7051L29.6055 52.3017L29.6757 52.2658L29.8898 52.1614L30.1922 52.1968L30.3503 52.3017L30.5272 52.4059L30.7411 52.6708L30.795 52.7407L30.9023 53.1625L30.9552 53.6732L31.275 53.9536L32.072 54.2015L32.1779 54.2712L32.266 54.3407L32.409 54.5002L32.4975 54.5708L32.5675 54.6235H32.8173L33.1369 54.5529L33.4002 54.5002H33.5969L33.8255 54.6405L34.1279 55.0439L34.3056 55.1842L34.3239 55.2033L34.5369 55.149L34.749 55.0966H34.8195L34.997 55.1143L35.1219 55.3962L35.1575 55.5013L35.05 55.6062L34.9597 55.6959L34.9071 55.9424L35.2264 56.2752L35.1381 56.329L35.1018 56.3483L35.0491 56.3655L34.9588 56.3989L34.8007 56.418L34.6426 56.3989L34.6607 56.4532L34.6778 56.5571L34.6952 56.6275L34.7307 56.6972V56.7165L34.6234 57.0143L34.5532 57.1573V57.1746L34.4468 57.3156L34.2887 57.1746L34.1814 57.2459L34.1274 57.368L34.1094 57.4556L34.092 57.5787L34.0746 57.7548L34.2679 58.0215L34.0385 58.5475L33.4878 59.4281L33.7905 59.3384H33.9307L34.0746 59.4081L34.0385 59.4794L33.9133 59.673L33.6649 60.1838L33.4872 60.4467L33.4171 60.4991L33.2575 60.5515L33.1346 60.6058L33.0971 60.6228L32.2631 61.5385L32.0697 61.7508L32.0869 62.0838V62.2429L32.1208 62.2952L32.3159 62.5744L32.6353 62.7869L32.6883 62.8043L32.9368 62.9285H32.9539L33.1506 62.9814L33.1335 63.1558V63.2261L33.077 63.4209L32.9896 63.738H33.007L33.1335 63.8084L33.2383 63.8781L33.3458 63.8971H33.3995L33.4704 63.8781L33.5581 63.8597L33.7548 63.561L33.772 63.5271L33.8948 63.424L33.9852 63.4576L34.038 63.4747L34.3203 63.7931L34.4642 63.8971L35.6867 63.8635L35.9873 63.9666H36.0063L37.0337 64L37.1412 63.9827L37.3171 63.9494L37.3528 63.6683L37.2643 63.5271L37.1213 63.424L37.0156 63.2461L36.9591 62.9659L37.1733 63.0191L37.3688 62.8238L37.4398 62.5592L37.4054 62.3673L37.3833 62.1694L37.5075 61.959L37.7031 61.8541L38.0193 61.8731L38.1787 61.6779L38.2691 61.4511V61.2741L38.1424 60.8165L38.1613 60.6755L38.2316 60.5507L38.2652 60.4632L37.8054 60.3064L37.4693 60.095L37.3627 59.9722L37.3273 59.797V59.4608L37.3982 59.2143L37.4856 59.0188L37.4504 58.8967L36.9372 58.9491L36.5648 58.8268L36.3688 58.8091L36.1741 58.8615L35.8014 59.0377L35.6065 59.071L35.5882 58.5081L35.3741 57.8754L35.2867 57.3646L35.6765 57.154L35.8733 57.2577L36.0494 57.4675L36.245 57.6468L36.509 57.613L36.6897 57.416L36.953 56.8557L37.1498 56.6428L38.443 55.6217L39.2402 55.1824L39.4543 54.903L39.1528 54.621L39.4708 54.3559L39.5774 54.2855L39.2576 54.1815L38.5498 54.6553L38.2128 54.4809L38.1051 54.0775L38.141 53.5845L38.2664 53.145L38.4075 52.8985L38.7099 52.5462L39.4369 51.2272L39.5774 51.0854L39.9306 50.8389L40.0729 50.6468L40.1441 50.3664L40.1802 49.8734L40.2706 49.6804L40.0562 49.4339L40.0738 49.0305L40.1811 48.5558L40.234 48.1163L40.164 47.7116L40.0562 46.972L39.4916 46.3203L38.5516 45.0368L38.497 44.755L38.5679 44.4413L38.726 43.929L37.7688 43.5255L37.4492 43.1221L37.6457 42.4875L37.5021 42.3463L37.1841 42.1177L37.0237 42.0464L36.6172 42.0654L36.4745 42.0128L36.4925 41.7662L36.8123 41.4514L36.8297 40.9915L36.6328 40.5515L36.3166 40.3064L36.3676 40.1112L36.5989 39.9556L37.1848 39.8483L37.451 39.7259L37.7334 39.5155L37.9283 39.4088L38.4947 39.2519L38.9557 39.3741L39.2926 39.4637L39.5243 39.5159L39.3281 39.2167L38.7973 38.6728L38.6196 38.5471L39.0977 38.2846L39.5578 38.4966L40.0196 38.8642L40.4978 39.0596L40.9407 38.97L41.0674 38.9872L41.1726 39.0948L41.3307 39.3741L41.4016 39.4453L41.9339 39.535L42.1643 39.6573L42.1986 40.0092L42.731 39.868L42.7854 39.6396L42.6788 39.1289L43.5822 39.3926H43.8645L44.0953 39.2878L44.7857 38.7609L46.186 38.286L47.444 38.2499L47.8166 38.0394L48.1554 37.6526L48.475 37.0699L48.5979 36.9482L48.7035 36.9299L48.7758 36.8602L48.793 36.5782L48.7575 36.4213L48.6873 36.3142L48.3338 35.9272L47.4121 35.4173L47.1113 35.3308L46.8983 35.1706L46.8622 34.8207L46.9692 34.5928L47.0944 34.5406H47.2342L47.4121 34.4509L47.519 34.3082L47.6407 34.0081L47.7134 33.8862L47.9619 33.6396L48.1748 33.5168L49.097 33.4137L49.2745 33.2557L49.5925 32.6219L49.8592 32.3418L50.4622 32.0766L50.7436 31.8317L51.0824 31.3032L51.2405 31.128L51.8607 30.7409L51.8959 30.6172L51.7733 30.4065L51.6484 29.9148V29.489L51.7539 29.1403L52.0364 28.4883L52.1945 27.6271L51.9996 27.0444L51.1499 25.6535L51.0596 25.6887L50.9722 25.8456L50.8473 25.9695L50.4932 26.0755L50.1205 26.0928L49.7305 26.0211L49.0028 25.7409L47.8871 25.584L47.516 25.4079L47.2319 25.0903L47.1434 24.7541L47.1244 24.3507L47.0707 23.9473L46.8947 23.5944L47.1244 23.5775L47.9058 23.1887L47.4433 22.8019L47.3383 22.6625V22.4655L47.4094 22.2741L47.427 22.0612L47.304 21.8147L47.7992 21.4639L48.2783 21.3214L49.625 21.3039L50.3681 21.5336L51.0596 21.4278L51.2744 21.5175L51.7701 21.6744L52.036 21.1829L52.1417 20.3738L52.2656 17.8736L52.3892 17.0988L52.6376 16.3956L53.7919 14.5831L54.0946 13.8088L54.2174 13.0693L54.2348 12.2423L54.1115 10.569L54.0569 10.3776L53.9873 10.2537L53.934 10.1127L53.9171 9.83146L53.9505 9.6385L54.1436 9.13044ZM35.5819 24.5251L35.4216 25.069L35.1566 25.5276L34.5889 26.3364L34.2892 26.6697L33.8282 27.0058L33.3492 27.2524L32.5336 27.4962L32.1074 27.8469L31.4687 28.6228L31.0089 29.6966L30.8492 29.8727L29.6278 30.1712L29.0248 30.1186L28.4035 29.8205L27.0911 28.8702L26.8068 28.8351L26.2055 28.9048L25.8686 28.8702L25.6016 28.766L24.8938 28.2719L24.6106 28.1688L23.9712 28.1165L23.7427 28.0468L22.8718 27.4993L22.9436 27.3944L23.031 27.3599L23.0502 27.3427V27.1665L23.2634 27.2008L23.3526 27.255L23.4583 27.3427V27.1665L23.6724 27.3247L23.7953 27.0782L23.9019 26.7259L24.0794 26.5515L24.2375 26.4991L24.3969 26.3563L24.5218 26.1647L24.6293 25.9182L24.3988 25.7958L24.2933 25.6024L24.2029 25.409L24.0433 25.2663L24.3276 25.1439L24.4333 25.1271L24.3631 24.8965L24.3285 24.8068L24.612 24.5087L24.5063 23.9485L24.046 22.9097L23.8495 22.2425L23.8323 21.6607L24.0462 21.2723L24.5419 21.1844L24.9144 21.3245L25.021 21.396L25.2705 21.6607L25.4654 21.6255L25.6795 21.5383L26.5477 20.9913L26.7772 20.7288L26.9547 20.1822L27.169 19.9906L27.6298 19.76L27.5769 19.532L27.5414 19.2165L27.4884 18.9337L27.3994 18.8122L26.6733 18.4944L26.4429 18.4774L26.248 18.5119L26.0886 18.5993L25.9474 18.7423L25.7523 18.8642L25.5757 18.7593L25.4155 18.5831L25.275 18.4935L25.0801 18.1762L25.0429 16.6986L24.9729 16.1894L24.2632 14.7631L23.9606 14.519L24.5478 14.2725L25.1492 13.9025L27.3838 11.933L28.5715 11.1764L28.7296 11.6683L28.8189 11.8826L28.9609 12.1291L29.1021 12.2156L29.1556 12.5159L29.1387 12.8144L29.0494 13.0946L29.1897 13.359L29.2453 13.7288L29.175 14.0986L28.9081 14.4124L29.105 14.7817L28.9783 14.9592L28.7493 15.0811L28.5901 15.346L28.5201 15.3812L28.4489 15.4507V15.6806L28.5038 15.8207L28.5919 15.8902L28.6987 15.9426L28.7877 16.0482L29.0883 16.2947L29.531 16.4186L30.4183 16.4556L30.594 16.5074L30.9312 16.7172L31.1444 16.7694L31.3741 16.7358L31.5168 16.6647L31.8166 16.4556L32.049 16.366H32.2628L33.2907 16.6645L33.4508 16.7882L32.9368 16.8918L32.5108 17.1921L32.1745 17.6506L31.8906 18.2109L31.7657 18.7936L31.6589 19.7597L31.6761 20.6562L31.9259 21.0433L32.1018 21.1856L32.4388 21.8187L32.5646 21.9951L32.7931 22.0645L34.6559 21.9597L34.9044 22.0121L35.0995 22.1549L35.2416 22.3992L35.347 22.786L35.5467 23.6419L35.5995 24.0677L35.5819 24.5251ZM32.8521 22.4115L32.4625 22.0081L32.214 22.2015L31.399 22.4646L31.1492 22.6416L30.529 23.2566L30.3885 23.2766L30.1042 23.2398L29.9987 23.3294L29.6599 24.1934L29.465 24.3335L28.9343 24.5464L28.2947 25.1246L27.2332 25.5466L26.6295 26.1461L25.7439 26.6206L25.3365 27.0058L25.088 27.1468L25.2303 27.4973L25.2832 27.8321L25.3715 28.0968L25.673 28.2537L26.3124 28.2194L26.5412 28.2537L27.5164 28.6237L27.9593 28.8702L28.8082 29.5394L29.18 29.6948L30.1218 29.6614L30.6173 29.4666L30.919 29.0632L31.0073 28.4821L30.9016 27.7766L31.2915 27.5652L31.4882 27.4931L31.7167 27.4758L31.9311 27.3709L32.2314 26.9319L32.4092 26.827L32.7818 26.7566L33.4555 26.4397L33.8621 26.3693L34.2532 26.2275L34.6252 25.8752L34.9258 25.4186L35.158 24.9792L35.3703 24.4162V23.9576L35.0862 22.9361L32.8521 22.4115ZM5.19702 43.2593L5.83392 43.2219L5.92313 43.1522L6.27839 42.7488L6.4733 42.7658L6.59638 42.8716L6.6856 42.9955L6.77481 43.0829L7.32317 43.312L7.55444 43.4872L7.64274 43.7701L7.78322 44.0686L8.08564 44.279L8.61751 44.4727L8.86595 44.4897L9.94822 44.279L10.1967 44.1916L10.427 44.0334L10.5693 43.8043L10.5338 43.435L10.2685 43.2407L9.68126 43.0838L9.54101 42.9606L9.4527 42.8364L9.38066 42.7145L8.83162 42.063L8.6008 41.8882L8.28123 41.8187L6.93403 41.7813L6.7398 41.6755L6.50853 41.6578L6.45591 41.6755L5.26794 41.853L4.75413 42.1865L4.71867 42.9077L4.91358 43.171L5.19702 43.2593ZM11.6665 50.0141L11.5782 50.1008L11.3297 50.4532L10.8507 50.418L10.7624 50.3837L10.6219 50.277L10.2832 50.2248L10.1219 50.1712L9.99904 50.0312L9.83959 49.6795L9.75128 49.5394L9.55637 49.4155L9.23657 49.2916L8.93506 49.2745L8.7939 49.4682L8.72366 49.9092L8.52695 50.136L7.14588 50.7187L6.80891 50.7714L6.27703 50.6121H5.70947L5.40706 50.489L4.98156 50.0489L3.59891 49.662L2.71268 49.0991L2.50128 49.1168L2.28718 49.2922L0 50.0318L0.0718186 50.5948L0.106601 50.8933L0.0537502 51.4585L0.0354579 51.9843L0.16103 52.3539L0.23127 52.5475L0.569817 52.9164L0.798603 53.0222L0.995318 53.1098L1.27876 53.145L1.42082 53.1623L1.98838 53.0744L2.09475 53.0574L2.11305 52.9677L2.13044 52.844L2.09475 52.5625L2.11305 52.4744L2.13044 52.2996L2.28853 52.1414H2.30592L2.50264 52.0874L2.7312 52.0701L3.29966 51.9482H3.33512L3.65492 51.9301L4.00838 51.9825L5.72822 52.5457L5.99495 52.7219L6.06496 53.0211L6.0478 53.2676L6.06496 53.4964L6.27839 53.7429L6.43784 53.8142L6.45523 53.8312L8.1211 53.8142L8.3352 53.7429L8.4409 53.7087L8.52831 53.6372L9.07938 53.2147L10.8329 52.3362L11.5253 51.7905L12.0755 51.0338L12.1283 50.8754L12.1453 50.8232L12.2521 50.3133V50.2241L11.9506 50.0475L11.6665 50.0141ZM19.9075 13.1757L20.3865 13.0356L20.7063 12.4191L21.3629 11.8752L21.7529 11.275L21.9478 10.9061L22.0361 10.6089H21.7545V10.4662L21.9968 10.2535V9.88391L21.8557 9.46168L21.6608 9.14635L21.7326 9.03967L21.75 8.93456L21.7136 8.81286L21.6608 8.66988L21.2179 8.65262L20.8644 9.23532L20.5277 9.93926L19.8524 10.5009L19.5154 11.0657L19.234 11.751L19.1437 12.3666L19.4086 12.963L19.9075 13.1757Z"
-          fill="var(--voorbeeld-logo-icon-color, currentColor)"
-        />
-        <path
-          d="M73.608 25.456V24.232H78.816V25.528L77.016 25.816L72.072 40.048H69.744L64.584 25.672L63.192 25.528V24.232H70.92V25.456L68.88 25.72L72.168 35.608L75.48 25.768L73.608 25.456ZM83.6436 40.36C82.3956 40.36 81.3156 40.104 80.4036 39.592C79.4916 39.064 78.7876 38.336 78.2916 37.408C77.7956 36.464 77.5476 35.376 77.5476 34.144C77.5476 32.912 77.8116 31.84 78.3396 30.928C78.8836 30.016 79.6196 29.304 80.5476 28.792C81.4756 28.28 82.5076 28.024 83.6436 28.024C84.7796 28.024 85.8036 28.272 86.7156 28.768C87.6436 29.264 88.3796 29.976 88.9236 30.904C89.4676 31.816 89.7396 32.896 89.7396 34.144C89.7396 35.392 89.4916 36.48 88.9956 37.408C88.4996 38.336 87.7956 39.064 86.8836 39.592C85.9716 40.104 84.8916 40.36 83.6436 40.36ZM83.6436 39.112C84.3156 39.112 84.8196 38.72 85.1556 37.936C85.4916 37.152 85.6596 35.912 85.6596 34.216C85.6596 32.504 85.4916 31.256 85.1556 30.472C84.8196 29.688 84.3156 29.296 83.6436 29.296C82.9556 29.296 82.4436 29.688 82.1076 30.472C81.7716 31.256 81.6036 32.504 81.6036 34.216C81.6036 35.912 81.7716 37.152 82.1076 37.936C82.4436 38.72 82.9556 39.112 83.6436 39.112ZM97.3779 40.36C96.1299 40.36 95.0499 40.104 94.1379 39.592C93.2259 39.064 92.5219 38.336 92.0259 37.408C91.5299 36.464 91.2819 35.376 91.2819 34.144C91.2819 32.912 91.5459 31.84 92.0739 30.928C92.6179 30.016 93.3539 29.304 94.2819 28.792C95.2099 28.28 96.2419 28.024 97.3779 28.024C98.5139 28.024 99.5379 28.272 100.45 28.768C101.378 29.264 102.114 29.976 102.658 30.904C103.202 31.816 103.474 32.896 103.474 34.144C103.474 35.392 103.226 36.48 102.73 37.408C102.234 38.336 101.53 39.064 100.618 39.592C99.7059 40.104 98.6259 40.36 97.3779 40.36ZM97.3779 39.112C98.0499 39.112 98.5539 38.72 98.8899 37.936C99.2259 37.152 99.3939 35.912 99.3939 34.216C99.3939 32.504 99.2259 31.256 98.8899 30.472C98.5539 29.688 98.0499 29.296 97.3779 29.296C96.6899 29.296 96.1779 29.688 95.8419 30.472C95.5059 31.256 95.3379 32.504 95.3379 34.216C95.3379 35.912 95.5059 37.152 95.8419 37.936C96.1779 38.72 96.6899 39.112 97.3779 39.112ZM104.804 40V38.872L106.124 38.608C106.14 38.016 106.148 37.384 106.148 36.712C106.148 36.04 106.148 35.456 106.148 34.96V33.544C106.148 33.096 106.14 32.728 106.124 32.44C106.124 32.152 106.116 31.888 106.1 31.648C106.1 31.392 106.092 31.104 106.076 30.784L104.612 30.568V29.632L109.34 28.024L109.772 28.288L109.988 31.12C110.324 30.096 110.796 29.328 111.404 28.816C112.012 28.288 112.612 28.024 113.204 28.024C113.732 28.024 114.188 28.176 114.572 28.48C114.956 28.768 115.196 29.208 115.292 29.8C115.276 30.424 115.108 30.912 114.788 31.264C114.468 31.6 114.068 31.768 113.588 31.768C112.916 31.768 112.3 31.392 111.74 30.64L111.596 30.448C111.244 30.784 110.916 31.208 110.612 31.72C110.324 32.232 110.116 32.768 109.988 33.328V34.96C109.988 35.456 109.988 36.024 109.988 36.664C109.988 37.304 109.996 37.904 110.012 38.464L111.908 38.872V40H104.804ZM115.423 40.024V38.896L116.695 38.632C116.711 38.056 116.719 37.432 116.719 36.76C116.719 36.072 116.719 35.48 116.719 34.984V24.856L115.231 24.664V23.656L120.199 22.576L120.583 22.816L120.487 26.176V29.56C121.495 28.536 122.655 28.024 123.967 28.024C124.847 28.024 125.639 28.256 126.343 28.72C127.047 29.168 127.607 29.848 128.023 30.76C128.455 31.656 128.671 32.768 128.671 34.096C128.671 35.392 128.431 36.512 127.951 37.456C127.471 38.384 126.839 39.104 126.055 39.616C125.287 40.112 124.447 40.36 123.535 40.36C122.223 40.36 121.167 39.888 120.367 38.944L119.959 40.336L115.423 40.024ZM122.215 30.112C121.911 30.112 121.623 30.176 121.351 30.304C121.079 30.432 120.815 30.6 120.559 30.808V37.672C121.055 38.12 121.599 38.344 122.191 38.344C122.895 38.344 123.471 38 123.919 37.312C124.383 36.624 124.615 35.576 124.615 34.168C124.615 32.744 124.391 31.712 123.943 31.072C123.511 30.432 122.935 30.112 122.215 30.112ZM136.163 29.272C135.715 29.272 135.299 29.552 134.915 30.112C134.547 30.656 134.339 31.664 134.291 33.136H136.331C136.907 33.136 137.291 33.024 137.483 32.8C137.691 32.56 137.795 32.168 137.795 31.624C137.795 30.824 137.635 30.232 137.315 29.848C136.995 29.464 136.611 29.272 136.163 29.272ZM136.211 40.36C135.043 40.36 134.011 40.12 133.115 39.64C132.219 39.144 131.515 38.44 131.003 37.528C130.491 36.6 130.235 35.496 130.235 34.216C130.235 32.872 130.531 31.744 131.123 30.832C131.731 29.904 132.507 29.208 133.451 28.744C134.395 28.264 135.379 28.024 136.403 28.024C137.459 28.024 138.355 28.248 139.091 28.696C139.843 29.144 140.411 29.752 140.795 30.52C141.179 31.272 141.371 32.136 141.371 33.112C141.371 33.576 141.323 34 141.227 34.384H134.291C134.371 35.648 134.715 36.584 135.323 37.192C135.947 37.784 136.675 38.08 137.507 38.08C138.163 38.08 138.723 37.96 139.187 37.72C139.651 37.464 140.059 37.152 140.411 36.784L141.251 37.552C140.755 38.496 140.083 39.2 139.235 39.664C138.387 40.128 137.379 40.36 136.211 40.36ZM148.679 29.272C148.231 29.272 147.815 29.552 147.431 30.112C147.063 30.656 146.855 31.664 146.807 33.136H148.847C149.423 33.136 149.807 33.024 149.999 32.8C150.207 32.56 150.311 32.168 150.311 31.624C150.311 30.824 150.151 30.232 149.831 29.848C149.511 29.464 149.127 29.272 148.679 29.272ZM148.727 40.36C147.559 40.36 146.527 40.12 145.631 39.64C144.735 39.144 144.031 38.44 143.519 37.528C143.007 36.6 142.751 35.496 142.751 34.216C142.751 32.872 143.047 31.744 143.639 30.832C144.247 29.904 145.023 29.208 145.967 28.744C146.911 28.264 147.895 28.024 148.919 28.024C149.975 28.024 150.871 28.248 151.607 28.696C152.359 29.144 152.927 29.752 153.311 30.52C153.695 31.272 153.887 32.136 153.887 33.112C153.887 33.576 153.839 34 153.743 34.384H146.807C146.887 35.648 147.231 36.584 147.839 37.192C148.463 37.784 149.191 38.08 150.023 38.08C150.679 38.08 151.239 37.96 151.703 37.72C152.167 37.464 152.575 37.152 152.927 36.784L153.767 37.552C153.271 38.496 152.599 39.2 151.751 39.664C150.903 40.128 149.895 40.36 148.727 40.36ZM154.796 40V38.872L156.116 38.584C156.132 37.976 156.14 37.368 156.14 36.76C156.156 36.152 156.164 35.552 156.164 34.96V24.904L154.676 24.664V23.656L159.716 22.576L160.1 22.816L160.004 26.176V34.96C160.004 35.568 160.004 36.176 160.004 36.784C160.02 37.392 160.036 38 160.052 38.608L161.372 38.872V40H154.796ZM166.682 34.264C166.682 35.576 166.914 36.576 167.378 37.264C167.842 37.936 168.442 38.272 169.178 38.272C169.722 38.272 170.218 38.088 170.666 37.72V30.568C170.234 30.232 169.762 30.064 169.25 30.064C168.514 30.064 167.898 30.408 167.402 31.096C166.922 31.784 166.682 32.84 166.682 34.264ZM170.978 40.264L170.786 38.896C169.858 39.872 168.73 40.36 167.402 40.36C166.506 40.36 165.69 40.128 164.954 39.664C164.234 39.2 163.666 38.52 163.25 37.624C162.834 36.712 162.626 35.608 162.626 34.312C162.626 33 162.874 31.88 163.37 30.952C163.882 30.008 164.538 29.288 165.338 28.792C166.154 28.28 167.01 28.024 167.906 28.024C168.53 28.024 169.074 28.136 169.538 28.36C170.002 28.568 170.426 28.88 170.81 29.296V24.904L169.13 24.664V23.656L174.194 22.576L174.578 22.816L174.482 26.176V38.632L175.658 38.872V40L170.978 40.264Z"
-          fill="var(--voorbeeld-logo-color, currentColor)"
-        />
-      </svg>
-    </Link>
-  );
-};
-
-export const Logo = () => {
-  return (
-    <svg
-      role="img"
-      aria-label="logo van gemeente Voorbeeld"
-      width="177"
-      height="64"
-      viewBox="0 0 177 64"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M32.8349 3.00852L33.5091 2.69139L33.8817 2.60354H35.3344L36.7346 2.28731L36.1671 1.987L32.764 2.10937L32.5526 2.02263L32.3213 2.21694L32.231 2.56746L32.4794 2.9182L32.8349 3.00852ZM26.2037 4.81847L28.1533 3.95742L28.6501 3.83505L29.2879 3.55334L29.87 3.53586L30.0664 3.43254L30.0312 3.3429L29.9955 3.16585L29.9612 3.07799L30.7935 3.04348L31.1835 2.90341L31.5205 2.44352L29.43 2.84827L29.2333 2.99149L29.002 3.18288L27.1062 3.5697H26.7885L26.5387 3.80009L25.7762 4.06342L25.6181 4.46683L25.7938 4.81824L26.2037 4.81847ZM39.1643 2.23555V2.30502L39.2533 2.4442L39.537 2.09233L41.9662 1.33573H42.1611V1.17884L41.7695 1.05715L40.3711 1.17884L39.9984 1.26669L39.4668 1.49619L39.1113 1.81152L39.2533 2.12909L39.1643 2.23555ZM45.7237 0.474226L45.7765 0.317347L45.884 0.281487L45.9177 0.230838H45.599L45.4561 0.352532L45.4224 0.474452L45.6345 0.581578L45.7237 0.474226ZM43.8808 0.335498L44.3792 0.507845L44.4131 0.419541L44.4501 0.386151L44.5021 0.350962H44.5734L43.7915 0L43.6131 0.0192735L43.6312 0.174585L43.8808 0.335498ZM22.3573 7.77611L22.676 7.31645L23.7937 6.24272L24.2002 6.08584L24.3768 5.98117H24.7486V5.8243L24.0426 5.71941L23.0834 6.2797L22.1617 7.12371L21.5587 7.89937L22.0011 8.09301L22.3573 7.77611ZM54.1436 9.13044L54.2685 8.74249L54.303 8.37562L54.2856 8.00583L54.2493 7.35253H54.0925L53.0613 7.0883L52.6356 6.84334L52.6024 6.45495L52.5322 6.26042V6.15643L52.5848 6.08584L52.6887 5.98117V5.8243H52.6015L52.4766 5.94599L52.3169 5.99866L52.1403 5.96504L52.0177 5.8243L51.7692 5.89377L51.2008 5.5762L50.7561 5.43523L50.6526 5.33259L50.5815 5.22568L50.5083 5.17234L50.4914 5.1199L50.1526 4.87337L50.0824 4.83976L49.6031 2.90453L49.4312 2.68669L49.1489 2.54415L47.7468 2.3169L44.4147 2.82609L42.3068 3.4424L41.9534 3.72165L41.8296 3.75639L41.6857 3.70349L41.4386 3.49305L41.2957 3.44083L38.5302 3.53048L37.1656 3.59973L36.2084 4.09166L35.8895 4.10959L35.3748 4.47916L33.1233 5.37763L32.6249 5.79964L31.9516 5.90475L31.6684 6.00964L31.42 6.27432L31.2061 6.59144L30.9753 6.85656L30.4626 7.08382L30.1602 7.34805L29.6633 7.9294L29.381 8.63088L28.9891 10.2156L28.7057 10.7956L28.4234 10.991L27.7666 11.1479L27.4997 11.3604L27.2885 11.6044L24.4482 14.1192L23.898 14.3658L23.3305 14.437L23.029 14.5437L22.428 15.0527L22.1082 15.2288L21.5571 15.2624L20.9733 15.0879L20.6526 14.6658L20.8647 13.9615L20.7592 13.7871L20.6187 13.9263L19.7487 13.7871L19.5906 13.9789L19.4849 14.4534L19.4321 14.9982L19.4147 15.4209L19.288 16.1773L18.687 17.7252L18.5438 18.4467L18.5095 18.7974L18.2611 19.9426L17.6571 24.3902L17.373 24.9888V25.1632L17.2481 25.8151L16.3964 28.3153L14.9603 30.883L13.24 33.3989L11.2195 35.8254L10.7421 36.1441L10.1919 36.2313L9.53695 36.0921L9.4466 36.2474L9.35627 36.4245L9.51662 36.8122L9.78177 37.7102L9.8371 38.0616L9.80096 38.1313L9.72892 38.221L9.65868 38.3433L9.62322 38.5195L9.65868 38.6605L9.72892 38.7501L9.81813 38.8212L10.1716 39.3129L10.28 39.5253L10.1013 39.6645L9.96019 39.7349L9.85449 39.7539L9.79894 39.7349L9.40912 39.5253L8.93077 39.4177L8.66404 39.4377L8.54095 39.5946L7.65562 39.8595L7.24909 40.139L7.37286 40.7564H7.51605L7.62242 40.7385L7.67527 40.6488L7.95848 40.5784H8.31284L8.6681 40.4744L8.93325 40.2799L9.34227 40.5601L9.67833 40.8774L9.90779 41.3528L9.99633 42.0894L10.191 42.3888L11.1658 42.6889L11.5195 42.9355L11.6615 43.0941L11.1823 43.128L10.9338 43.232L10.7563 43.4274L10.7224 43.6199V43.8664L10.7563 44.13L10.8272 44.3245L10.9338 44.4467L11.075 44.5531L11.2179 44.6056L11.8375 44.6246L12.0143 44.7974L12.0852 45.2212L11.7464 44.9417L11.3749 44.835L10.1169 44.7638L9.74428 44.8687L8.99897 45.2212L10.1167 46.2571L10.4521 46.4532L10.666 46.4868L11.1089 46.4341L11.3218 46.4532L11.9068 46.7151L12.5462 46.8212L12.936 47.048L13.2375 47.4178L13.3604 47.9121L13.1673 48.422L12.742 48.5437L12.2819 48.5269L11.9978 48.6134L11.7128 48.474L10.9858 48.5789L10.6513 48.474L10.4374 48.193L10.1531 47.4178L9.96109 47.0641L9.37547 46.5536L8.68504 46.2732L7.33785 46.1331V45.9922L7.42616 45.9025L7.67459 45.7456L7.81778 45.6936L7.69289 45.359L7.49798 45.0789L7.26761 44.8673L7.05531 44.7425L6.73641 44.7089L2.94214 45.1123L1.48925 46.0108L1.34629 46.2042L1.39914 46.5737L1.55926 46.8203L2.02113 47.225L2.19706 47.4866L2.39197 47.8389L2.60495 48.156L2.85338 48.2992L3.65131 48.4731L3.97088 48.4554L4.55583 48.2432L4.76813 48.2609L5.3183 48.7907L6.16998 49.212L6.5598 49.6154L6.70118 49.6851L6.85927 49.6672L7.03996 49.5442L7.55263 49.4917L7.76584 49.3861L7.92528 49.1598L8.35056 48.3693L8.54366 48.2436L9.13087 48.2956L9.36237 48.4023L10.1429 49.0706L10.427 49.2461L10.7288 49.3335L12.022 49.3525L12.34 49.2629L12.5003 49.0513L12.6406 49.1564L12.676 49.3308V49.5614L12.7115 49.8409L12.8183 50.1053L12.8727 50.1748L13.2793 50.1941L13.4387 50.2826L13.7931 50.4742L13.9512 50.5269L14.0569 50.5623L14.608 50.5975H14.7136L14.9804 50.4572L14.9987 50.3675L15.0515 50.0672L15.0332 50.0338L14.9614 49.8568L14.6436 49.489L14.608 49.4366L14.5379 49.2956V49.0668L14.6283 48.8904L14.6619 48.7335L14.486 48.5392L15.2313 48.1172L15.7094 47.9426L15.9764 47.8359L16.3472 47.8026H16.3998L16.5767 47.8359L16.6824 48.0329L16.665 48.3124V48.4007L16.5419 48.6811L16.5229 48.786L16.5055 48.8906L16.5932 48.9603L16.7707 49.0845L16.8581 49.1031H16.9484L17.055 49.0845L17.3574 49.0307L17.782 49.1387H17.8349L18.0307 49.1197L18.2256 49.0484L18.3686 48.9435L19.4673 47.6082L19.929 47.3617L20.1221 47.3803L20.2305 47.4139L20.3716 47.4666L20.4419 47.5009L20.6176 47.6266L20.7608 47.818L20.7971 48.1542L20.7413 48.3664L20.7079 48.4886L20.6704 48.7178L20.653 48.8227L20.7079 49.0323L20.7608 49.2095L20.1404 48.9276L19.9823 48.9467L19.8931 49.2125L20.1943 49.3335L21.1881 49.3687L21.3136 49.4023L21.4548 49.4384L21.7199 49.5623L22.0034 49.6842L22.3049 49.5966L22.465 49.4191L22.6602 49.227L22.8743 48.9276L22.9797 48.7875L23.0674 48.6439L23.1941 48.4343L23.2469 48.2934L23.2995 48.1685L23.3361 48.0282L23.3551 47.957L23.4425 47.833L23.6559 47.9758L23.6749 48.011L23.6986 48.0253L24.123 48.3046L24.2287 48.4984L24.4062 48.8346L24.3534 49.2044L24.1953 49.5551L24.177 49.7297L24.1424 50.0276L24.2298 50.2402L24.7079 50.7665L24.78 50.9612L24.9202 51.4543L24.9751 51.576V51.5952H24.9923L25.2247 51.7187L25.7721 51.6649H25.8286L26.0215 51.6827L26.2713 52.0011V52.7907L26.5197 52.8624L27.0509 52.7387L28.7179 52.81L29.0183 52.7051L29.6055 52.3017L29.6757 52.2658L29.8898 52.1614L30.1922 52.1968L30.3503 52.3017L30.5272 52.4059L30.7411 52.6708L30.795 52.7407L30.9023 53.1625L30.9552 53.6732L31.275 53.9536L32.072 54.2015L32.1779 54.2712L32.266 54.3407L32.409 54.5002L32.4975 54.5708L32.5675 54.6235H32.8173L33.1369 54.5529L33.4002 54.5002H33.5969L33.8255 54.6405L34.1279 55.0439L34.3056 55.1842L34.3239 55.2033L34.5369 55.149L34.749 55.0966H34.8195L34.997 55.1143L35.1219 55.3962L35.1575 55.5013L35.05 55.6062L34.9597 55.6959L34.9071 55.9424L35.2264 56.2752L35.1381 56.329L35.1018 56.3483L35.0491 56.3655L34.9588 56.3989L34.8007 56.418L34.6426 56.3989L34.6607 56.4532L34.6778 56.5571L34.6952 56.6275L34.7307 56.6972V56.7165L34.6234 57.0143L34.5532 57.1573V57.1746L34.4468 57.3156L34.2887 57.1746L34.1814 57.2459L34.1274 57.368L34.1094 57.4556L34.092 57.5787L34.0746 57.7548L34.2679 58.0215L34.0385 58.5475L33.4878 59.4281L33.7905 59.3384H33.9307L34.0746 59.4081L34.0385 59.4794L33.9133 59.673L33.6649 60.1838L33.4872 60.4467L33.4171 60.4991L33.2575 60.5515L33.1346 60.6058L33.0971 60.6228L32.2631 61.5385L32.0697 61.7508L32.0869 62.0838V62.2429L32.1208 62.2952L32.3159 62.5744L32.6353 62.7869L32.6883 62.8043L32.9368 62.9285H32.9539L33.1506 62.9814L33.1335 63.1558V63.2261L33.077 63.4209L32.9896 63.738H33.007L33.1335 63.8084L33.2383 63.8781L33.3458 63.8971H33.3995L33.4704 63.8781L33.5581 63.8597L33.7548 63.561L33.772 63.5271L33.8948 63.424L33.9852 63.4576L34.038 63.4747L34.3203 63.7931L34.4642 63.8971L35.6867 63.8635L35.9873 63.9666H36.0063L37.0337 64L37.1412 63.9827L37.3171 63.9494L37.3528 63.6683L37.2643 63.5271L37.1213 63.424L37.0156 63.2461L36.9591 62.9659L37.1733 63.0191L37.3688 62.8238L37.4398 62.5592L37.4054 62.3673L37.3833 62.1694L37.5075 61.959L37.7031 61.8541L38.0193 61.8731L38.1787 61.6779L38.2691 61.4511V61.2741L38.1424 60.8165L38.1613 60.6755L38.2316 60.5507L38.2652 60.4632L37.8054 60.3064L37.4693 60.095L37.3627 59.9722L37.3273 59.797V59.4608L37.3982 59.2143L37.4856 59.0188L37.4504 58.8967L36.9372 58.9491L36.5648 58.8268L36.3688 58.8091L36.1741 58.8615L35.8014 59.0377L35.6065 59.071L35.5882 58.5081L35.3741 57.8754L35.2867 57.3646L35.6765 57.154L35.8733 57.2577L36.0494 57.4675L36.245 57.6468L36.509 57.613L36.6897 57.416L36.953 56.8557L37.1498 56.6428L38.443 55.6217L39.2402 55.1824L39.4543 54.903L39.1528 54.621L39.4708 54.3559L39.5774 54.2855L39.2576 54.1815L38.5498 54.6553L38.2128 54.4809L38.1051 54.0775L38.141 53.5845L38.2664 53.145L38.4075 52.8985L38.7099 52.5462L39.4369 51.2272L39.5774 51.0854L39.9306 50.8389L40.0729 50.6468L40.1441 50.3664L40.1802 49.8734L40.2706 49.6804L40.0562 49.4339L40.0738 49.0305L40.1811 48.5558L40.234 48.1163L40.164 47.7116L40.0562 46.972L39.4916 46.3203L38.5516 45.0368L38.497 44.755L38.5679 44.4413L38.726 43.929L37.7688 43.5255L37.4492 43.1221L37.6457 42.4875L37.5021 42.3463L37.1841 42.1177L37.0237 42.0464L36.6172 42.0654L36.4745 42.0128L36.4925 41.7662L36.8123 41.4514L36.8297 40.9915L36.6328 40.5515L36.3166 40.3064L36.3676 40.1112L36.5989 39.9556L37.1848 39.8483L37.451 39.7259L37.7334 39.5155L37.9283 39.4088L38.4947 39.2519L38.9557 39.3741L39.2926 39.4637L39.5243 39.5159L39.3281 39.2167L38.7973 38.6728L38.6196 38.5471L39.0977 38.2846L39.5578 38.4966L40.0196 38.8642L40.4978 39.0596L40.9407 38.97L41.0674 38.9872L41.1726 39.0948L41.3307 39.3741L41.4016 39.4453L41.9339 39.535L42.1643 39.6573L42.1986 40.0092L42.731 39.868L42.7854 39.6396L42.6788 39.1289L43.5822 39.3926H43.8645L44.0953 39.2878L44.7857 38.7609L46.186 38.286L47.444 38.2499L47.8166 38.0394L48.1554 37.6526L48.475 37.0699L48.5979 36.9482L48.7035 36.9299L48.7758 36.8602L48.793 36.5782L48.7575 36.4213L48.6873 36.3142L48.3338 35.9272L47.4121 35.4173L47.1113 35.3308L46.8983 35.1706L46.8622 34.8207L46.9692 34.5928L47.0944 34.5406H47.2342L47.4121 34.4509L47.519 34.3082L47.6407 34.0081L47.7134 33.8862L47.9619 33.6396L48.1748 33.5168L49.097 33.4137L49.2745 33.2557L49.5925 32.6219L49.8592 32.3418L50.4622 32.0766L50.7436 31.8317L51.0824 31.3032L51.2405 31.128L51.8607 30.7409L51.8959 30.6172L51.7733 30.4065L51.6484 29.9148V29.489L51.7539 29.1403L52.0364 28.4883L52.1945 27.6271L51.9996 27.0444L51.1499 25.6535L51.0596 25.6887L50.9722 25.8456L50.8473 25.9695L50.4932 26.0755L50.1205 26.0928L49.7305 26.0211L49.0028 25.7409L47.8871 25.584L47.516 25.4079L47.2319 25.0903L47.1434 24.7541L47.1244 24.3507L47.0707 23.9473L46.8947 23.5944L47.1244 23.5775L47.9058 23.1887L47.4433 22.8019L47.3383 22.6625V22.4655L47.4094 22.2741L47.427 22.0612L47.304 21.8147L47.7992 21.4639L48.2783 21.3214L49.625 21.3039L50.3681 21.5336L51.0596 21.4278L51.2744 21.5175L51.7701 21.6744L52.036 21.1829L52.1417 20.3738L52.2656 17.8736L52.3892 17.0988L52.6376 16.3956L53.7919 14.5831L54.0946 13.8088L54.2174 13.0693L54.2348 12.2423L54.1115 10.569L54.0569 10.3776L53.9873 10.2537L53.934 10.1127L53.9171 9.83146L53.9505 9.6385L54.1436 9.13044ZM35.5819 24.5251L35.4216 25.069L35.1566 25.5276L34.5889 26.3364L34.2892 26.6697L33.8282 27.0058L33.3492 27.2524L32.5336 27.4962L32.1074 27.8469L31.4687 28.6228L31.0089 29.6966L30.8492 29.8727L29.6278 30.1712L29.0248 30.1186L28.4035 29.8205L27.0911 28.8702L26.8068 28.8351L26.2055 28.9048L25.8686 28.8702L25.6016 28.766L24.8938 28.2719L24.6106 28.1688L23.9712 28.1165L23.7427 28.0468L22.8718 27.4993L22.9436 27.3944L23.031 27.3599L23.0502 27.3427V27.1665L23.2634 27.2008L23.3526 27.255L23.4583 27.3427V27.1665L23.6724 27.3247L23.7953 27.0782L23.9019 26.7259L24.0794 26.5515L24.2375 26.4991L24.3969 26.3563L24.5218 26.1647L24.6293 25.9182L24.3988 25.7958L24.2933 25.6024L24.2029 25.409L24.0433 25.2663L24.3276 25.1439L24.4333 25.1271L24.3631 24.8965L24.3285 24.8068L24.612 24.5087L24.5063 23.9485L24.046 22.9097L23.8495 22.2425L23.8323 21.6607L24.0462 21.2723L24.5419 21.1844L24.9144 21.3245L25.021 21.396L25.2705 21.6607L25.4654 21.6255L25.6795 21.5383L26.5477 20.9913L26.7772 20.7288L26.9547 20.1822L27.169 19.9906L27.6298 19.76L27.5769 19.532L27.5414 19.2165L27.4884 18.9337L27.3994 18.8122L26.6733 18.4944L26.4429 18.4774L26.248 18.5119L26.0886 18.5993L25.9474 18.7423L25.7523 18.8642L25.5757 18.7593L25.4155 18.5831L25.275 18.4935L25.0801 18.1762L25.0429 16.6986L24.9729 16.1894L24.2632 14.7631L23.9606 14.519L24.5478 14.2725L25.1492 13.9025L27.3838 11.933L28.5715 11.1764L28.7296 11.6683L28.8189 11.8826L28.9609 12.1291L29.1021 12.2156L29.1556 12.5159L29.1387 12.8144L29.0494 13.0946L29.1897 13.359L29.2453 13.7288L29.175 14.0986L28.9081 14.4124L29.105 14.7817L28.9783 14.9592L28.7493 15.0811L28.5901 15.346L28.5201 15.3812L28.4489 15.4507V15.6806L28.5038 15.8207L28.5919 15.8902L28.6987 15.9426L28.7877 16.0482L29.0883 16.2947L29.531 16.4186L30.4183 16.4556L30.594 16.5074L30.9312 16.7172L31.1444 16.7694L31.3741 16.7358L31.5168 16.6647L31.8166 16.4556L32.049 16.366H32.2628L33.2907 16.6645L33.4508 16.7882L32.9368 16.8918L32.5108 17.1921L32.1745 17.6506L31.8906 18.2109L31.7657 18.7936L31.6589 19.7597L31.6761 20.6562L31.9259 21.0433L32.1018 21.1856L32.4388 21.8187L32.5646 21.9951L32.7931 22.0645L34.6559 21.9597L34.9044 22.0121L35.0995 22.1549L35.2416 22.3992L35.347 22.786L35.5467 23.6419L35.5995 24.0677L35.5819 24.5251ZM32.8521 22.4115L32.4625 22.0081L32.214 22.2015L31.399 22.4646L31.1492 22.6416L30.529 23.2566L30.3885 23.2766L30.1042 23.2398L29.9987 23.3294L29.6599 24.1934L29.465 24.3335L28.9343 24.5464L28.2947 25.1246L27.2332 25.5466L26.6295 26.1461L25.7439 26.6206L25.3365 27.0058L25.088 27.1468L25.2303 27.4973L25.2832 27.8321L25.3715 28.0968L25.673 28.2537L26.3124 28.2194L26.5412 28.2537L27.5164 28.6237L27.9593 28.8702L28.8082 29.5394L29.18 29.6948L30.1218 29.6614L30.6173 29.4666L30.919 29.0632L31.0073 28.4821L30.9016 27.7766L31.2915 27.5652L31.4882 27.4931L31.7167 27.4758L31.9311 27.3709L32.2314 26.9319L32.4092 26.827L32.7818 26.7566L33.4555 26.4397L33.8621 26.3693L34.2532 26.2275L34.6252 25.8752L34.9258 25.4186L35.158 24.9792L35.3703 24.4162V23.9576L35.0862 22.9361L32.8521 22.4115ZM5.19702 43.2593L5.83392 43.2219L5.92313 43.1522L6.27839 42.7488L6.4733 42.7658L6.59638 42.8716L6.6856 42.9955L6.77481 43.0829L7.32317 43.312L7.55444 43.4872L7.64274 43.7701L7.78322 44.0686L8.08564 44.279L8.61751 44.4727L8.86595 44.4897L9.94822 44.279L10.1967 44.1916L10.427 44.0334L10.5693 43.8043L10.5338 43.435L10.2685 43.2407L9.68126 43.0838L9.54101 42.9606L9.4527 42.8364L9.38066 42.7145L8.83162 42.063L8.6008 41.8882L8.28123 41.8187L6.93403 41.7813L6.7398 41.6755L6.50853 41.6578L6.45591 41.6755L5.26794 41.853L4.75413 42.1865L4.71867 42.9077L4.91358 43.171L5.19702 43.2593ZM11.6665 50.0141L11.5782 50.1008L11.3297 50.4532L10.8507 50.418L10.7624 50.3837L10.6219 50.277L10.2832 50.2248L10.1219 50.1712L9.99904 50.0312L9.83959 49.6795L9.75128 49.5394L9.55637 49.4155L9.23657 49.2916L8.93506 49.2745L8.7939 49.4682L8.72366 49.9092L8.52695 50.136L7.14588 50.7187L6.80891 50.7714L6.27703 50.6121H5.70947L5.40706 50.489L4.98156 50.0489L3.59891 49.662L2.71268 49.0991L2.50128 49.1168L2.28718 49.2922L0 50.0318L0.0718186 50.5948L0.106601 50.8933L0.0537502 51.4585L0.0354579 51.9843L0.16103 52.3539L0.23127 52.5475L0.569817 52.9164L0.798603 53.0222L0.995318 53.1098L1.27876 53.145L1.42082 53.1623L1.98838 53.0744L2.09475 53.0574L2.11305 52.9677L2.13044 52.844L2.09475 52.5625L2.11305 52.4744L2.13044 52.2996L2.28853 52.1414H2.30592L2.50264 52.0874L2.7312 52.0701L3.29966 51.9482H3.33512L3.65492 51.9301L4.00838 51.9825L5.72822 52.5457L5.99495 52.7219L6.06496 53.0211L6.0478 53.2676L6.06496 53.4964L6.27839 53.7429L6.43784 53.8142L6.45523 53.8312L8.1211 53.8142L8.3352 53.7429L8.4409 53.7087L8.52831 53.6372L9.07938 53.2147L10.8329 52.3362L11.5253 51.7905L12.0755 51.0338L12.1283 50.8754L12.1453 50.8232L12.2521 50.3133V50.2241L11.9506 50.0475L11.6665 50.0141ZM19.9075 13.1757L20.3865 13.0356L20.7063 12.4191L21.3629 11.8752L21.7529 11.275L21.9478 10.9061L22.0361 10.6089H21.7545V10.4662L21.9968 10.2535V9.88391L21.8557 9.46168L21.6608 9.14635L21.7326 9.03967L21.75 8.93456L21.7136 8.81286L21.6608 8.66988L21.2179 8.65262L20.8644 9.23532L20.5277 9.93926L19.8524 10.5009L19.5154 11.0657L19.234 11.751L19.1437 12.3666L19.4086 12.963L19.9075 13.1757Z"
-        fill="currentColor"
-      />
-      <path
-        d="M73.608 25.456V24.232H78.816V25.528L77.016 25.816L72.072 40.048H69.744L64.584 25.672L63.192 25.528V24.232H70.92V25.456L68.88 25.72L72.168 35.608L75.48 25.768L73.608 25.456ZM83.6436 40.36C82.3956 40.36 81.3156 40.104 80.4036 39.592C79.4916 39.064 78.7876 38.336 78.2916 37.408C77.7956 36.464 77.5476 35.376 77.5476 34.144C77.5476 32.912 77.8116 31.84 78.3396 30.928C78.8836 30.016 79.6196 29.304 80.5476 28.792C81.4756 28.28 82.5076 28.024 83.6436 28.024C84.7796 28.024 85.8036 28.272 86.7156 28.768C87.6436 29.264 88.3796 29.976 88.9236 30.904C89.4676 31.816 89.7396 32.896 89.7396 34.144C89.7396 35.392 89.4916 36.48 88.9956 37.408C88.4996 38.336 87.7956 39.064 86.8836 39.592C85.9716 40.104 84.8916 40.36 83.6436 40.36ZM83.6436 39.112C84.3156 39.112 84.8196 38.72 85.1556 37.936C85.4916 37.152 85.6596 35.912 85.6596 34.216C85.6596 32.504 85.4916 31.256 85.1556 30.472C84.8196 29.688 84.3156 29.296 83.6436 29.296C82.9556 29.296 82.4436 29.688 82.1076 30.472C81.7716 31.256 81.6036 32.504 81.6036 34.216C81.6036 35.912 81.7716 37.152 82.1076 37.936C82.4436 38.72 82.9556 39.112 83.6436 39.112ZM97.3779 40.36C96.1299 40.36 95.0499 40.104 94.1379 39.592C93.2259 39.064 92.5219 38.336 92.0259 37.408C91.5299 36.464 91.2819 35.376 91.2819 34.144C91.2819 32.912 91.5459 31.84 92.0739 30.928C92.6179 30.016 93.3539 29.304 94.2819 28.792C95.2099 28.28 96.2419 28.024 97.3779 28.024C98.5139 28.024 99.5379 28.272 100.45 28.768C101.378 29.264 102.114 29.976 102.658 30.904C103.202 31.816 103.474 32.896 103.474 34.144C103.474 35.392 103.226 36.48 102.73 37.408C102.234 38.336 101.53 39.064 100.618 39.592C99.7059 40.104 98.6259 40.36 97.3779 40.36ZM97.3779 39.112C98.0499 39.112 98.5539 38.72 98.8899 37.936C99.2259 37.152 99.3939 35.912 99.3939 34.216C99.3939 32.504 99.2259 31.256 98.8899 30.472C98.5539 29.688 98.0499 29.296 97.3779 29.296C96.6899 29.296 96.1779 29.688 95.8419 30.472C95.5059 31.256 95.3379 32.504 95.3379 34.216C95.3379 35.912 95.5059 37.152 95.8419 37.936C96.1779 38.72 96.6899 39.112 97.3779 39.112ZM104.804 40V38.872L106.124 38.608C106.14 38.016 106.148 37.384 106.148 36.712C106.148 36.04 106.148 35.456 106.148 34.96V33.544C106.148 33.096 106.14 32.728 106.124 32.44C106.124 32.152 106.116 31.888 106.1 31.648C106.1 31.392 106.092 31.104 106.076 30.784L104.612 30.568V29.632L109.34 28.024L109.772 28.288L109.988 31.12C110.324 30.096 110.796 29.328 111.404 28.816C112.012 28.288 112.612 28.024 113.204 28.024C113.732 28.024 114.188 28.176 114.572 28.48C114.956 28.768 115.196 29.208 115.292 29.8C115.276 30.424 115.108 30.912 114.788 31.264C114.468 31.6 114.068 31.768 113.588 31.768C112.916 31.768 112.3 31.392 111.74 30.64L111.596 30.448C111.244 30.784 110.916 31.208 110.612 31.72C110.324 32.232 110.116 32.768 109.988 33.328V34.96C109.988 35.456 109.988 36.024 109.988 36.664C109.988 37.304 109.996 37.904 110.012 38.464L111.908 38.872V40H104.804ZM115.423 40.024V38.896L116.695 38.632C116.711 38.056 116.719 37.432 116.719 36.76C116.719 36.072 116.719 35.48 116.719 34.984V24.856L115.231 24.664V23.656L120.199 22.576L120.583 22.816L120.487 26.176V29.56C121.495 28.536 122.655 28.024 123.967 28.024C124.847 28.024 125.639 28.256 126.343 28.72C127.047 29.168 127.607 29.848 128.023 30.76C128.455 31.656 128.671 32.768 128.671 34.096C128.671 35.392 128.431 36.512 127.951 37.456C127.471 38.384 126.839 39.104 126.055 39.616C125.287 40.112 124.447 40.36 123.535 40.36C122.223 40.36 121.167 39.888 120.367 38.944L119.959 40.336L115.423 40.024ZM122.215 30.112C121.911 30.112 121.623 30.176 121.351 30.304C121.079 30.432 120.815 30.6 120.559 30.808V37.672C121.055 38.12 121.599 38.344 122.191 38.344C122.895 38.344 123.471 38 123.919 37.312C124.383 36.624 124.615 35.576 124.615 34.168C124.615 32.744 124.391 31.712 123.943 31.072C123.511 30.432 122.935 30.112 122.215 30.112ZM136.163 29.272C135.715 29.272 135.299 29.552 134.915 30.112C134.547 30.656 134.339 31.664 134.291 33.136H136.331C136.907 33.136 137.291 33.024 137.483 32.8C137.691 32.56 137.795 32.168 137.795 31.624C137.795 30.824 137.635 30.232 137.315 29.848C136.995 29.464 136.611 29.272 136.163 29.272ZM136.211 40.36C135.043 40.36 134.011 40.12 133.115 39.64C132.219 39.144 131.515 38.44 131.003 37.528C130.491 36.6 130.235 35.496 130.235 34.216C130.235 32.872 130.531 31.744 131.123 30.832C131.731 29.904 132.507 29.208 133.451 28.744C134.395 28.264 135.379 28.024 136.403 28.024C137.459 28.024 138.355 28.248 139.091 28.696C139.843 29.144 140.411 29.752 140.795 30.52C141.179 31.272 141.371 32.136 141.371 33.112C141.371 33.576 141.323 34 141.227 34.384H134.291C134.371 35.648 134.715 36.584 135.323 37.192C135.947 37.784 136.675 38.08 137.507 38.08C138.163 38.08 138.723 37.96 139.187 37.72C139.651 37.464 140.059 37.152 140.411 36.784L141.251 37.552C140.755 38.496 140.083 39.2 139.235 39.664C138.387 40.128 137.379 40.36 136.211 40.36ZM148.679 29.272C148.231 29.272 147.815 29.552 147.431 30.112C147.063 30.656 146.855 31.664 146.807 33.136H148.847C149.423 33.136 149.807 33.024 149.999 32.8C150.207 32.56 150.311 32.168 150.311 31.624C150.311 30.824 150.151 30.232 149.831 29.848C149.511 29.464 149.127 29.272 148.679 29.272ZM148.727 40.36C147.559 40.36 146.527 40.12 145.631 39.64C144.735 39.144 144.031 38.44 143.519 37.528C143.007 36.6 142.751 35.496 142.751 34.216C142.751 32.872 143.047 31.744 143.639 30.832C144.247 29.904 145.023 29.208 145.967 28.744C146.911 28.264 147.895 28.024 148.919 28.024C149.975 28.024 150.871 28.248 151.607 28.696C152.359 29.144 152.927 29.752 153.311 30.52C153.695 31.272 153.887 32.136 153.887 33.112C153.887 33.576 153.839 34 153.743 34.384H146.807C146.887 35.648 147.231 36.584 147.839 37.192C148.463 37.784 149.191 38.08 150.023 38.08C150.679 38.08 151.239 37.96 151.703 37.72C152.167 37.464 152.575 37.152 152.927 36.784L153.767 37.552C153.271 38.496 152.599 39.2 151.751 39.664C150.903 40.128 149.895 40.36 148.727 40.36ZM154.796 40V38.872L156.116 38.584C156.132 37.976 156.14 37.368 156.14 36.76C156.156 36.152 156.164 35.552 156.164 34.96V24.904L154.676 24.664V23.656L159.716 22.576L160.1 22.816L160.004 26.176V34.96C160.004 35.568 160.004 36.176 160.004 36.784C160.02 37.392 160.036 38 160.052 38.608L161.372 38.872V40H154.796ZM166.682 34.264C166.682 35.576 166.914 36.576 167.378 37.264C167.842 37.936 168.442 38.272 169.178 38.272C169.722 38.272 170.218 38.088 170.666 37.72V30.568C170.234 30.232 169.762 30.064 169.25 30.064C168.514 30.064 167.898 30.408 167.402 31.096C166.922 31.784 166.682 32.84 166.682 34.264ZM170.978 40.264L170.786 38.896C169.858 39.872 168.73 40.36 167.402 40.36C166.506 40.36 165.69 40.128 164.954 39.664C164.234 39.2 163.666 38.52 163.25 37.624C162.834 36.712 162.626 35.608 162.626 34.312C162.626 33 162.874 31.88 163.37 30.952C163.882 30.008 164.538 29.288 165.338 28.792C166.154 28.28 167.01 28.024 167.906 28.024C168.53 28.024 169.074 28.136 169.538 28.36C170.002 28.568 170.426 28.88 170.81 29.296V24.904L169.13 24.664V23.656L174.194 22.576L174.578 22.816L174.482 26.176V38.632L175.658 38.872V40L170.978 40.264Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-};
-
-export const DenHaagLogo = () => {
-  return (
-    <svg
-      className="denhaag-header__logo"
-      xmlns="http://www.w3.org/2000/svg"
-      width="153"
-      height="48"
-      viewBox="0 0 153 48"
-      fill="none"
-    >
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M133.269 22.8104C133.269 23.7002 132.638 24.6354 131.676 24.6354C131.105 24.6354 130.759 24.2431 130.759 23.7305C130.759 23.0818 131.3 22.5842 133.269 22.5842V22.8104ZM135.163 20.9099C135.163 19.2508 134.441 18.2705 132.142 18.2705C131.225 18.2705 130.188 18.4967 129.361 18.7983L129.647 20.2161C130.368 19.9294 131.24 19.7184 131.916 19.7184C132.923 19.7184 133.269 20.1407 133.269 21.0305V21.3472C130.654 21.3472 128.94 22.1918 128.94 23.9566C128.94 25.1782 129.737 26.0831 131.074 26.0831C132.202 26.0831 133.118 25.4798 133.479 24.6353L133.494 24.6502C133.434 25.0425 133.419 25.4949 133.419 25.9474C134.381 25.9775 135.523 25.9323 136.47 25.8116V24.6502C136.034 24.6202 135.523 24.5599 135.163 24.4994V20.9099ZM123.807 16.9583C124.153 17.0337 124.558 17.0942 124.919 17.1244V19.8392H120.546V17.1244C120.906 17.0942 121.312 17.0337 121.658 16.9583V15.7518H117.479V16.9281C117.87 17.0188 118.336 17.0942 118.787 17.1244V24.5599C118.336 24.5901 117.87 24.6655 117.479 24.7559V25.9324H121.658V24.7258C121.312 24.6503 120.906 24.5901 120.546 24.5599V21.4983H124.919V24.5599C124.558 24.5901 124.153 24.6503 123.807 24.7258V25.9324H127.985V24.7559C127.594 24.6655 127.129 24.5901 126.678 24.5599V17.1244C127.129 17.0942 127.594 17.0188 127.985 16.9281V15.7518H123.807V16.9583ZM150.402 22.0107C150.402 23.323 149.786 24.5145 148.749 24.5145C147.862 24.5145 147.441 23.7905 147.441 22.3879C147.441 20.5026 148.238 19.7938 149.365 19.7938C149.741 19.7938 150.071 19.839 150.402 19.9144V22.0107ZM149.545 18.3007C146.915 18.3007 145.472 20.0352 145.472 22.5689C145.472 24.8615 146.464 26.0831 148.253 26.0831C149.245 26.0831 150.086 25.5402 150.477 24.6655L150.507 24.6804C150.432 25.0727 150.417 25.4045 150.417 25.7815V26.2492C150.417 27.5311 149.741 28.1042 148.313 28.1042C147.576 28.1042 146.765 27.9685 146.103 27.7272L145.923 29.4314C146.554 29.5822 147.366 29.6879 148.207 29.6879C151.048 29.6879 152.265 28.3455 152.265 26.0682V18.527C151.454 18.3911 150.582 18.3007 149.545 18.3007ZM141.55 22.8104C141.55 23.7002 140.918 24.6354 139.957 24.6354C139.385 24.6354 139.04 24.2431 139.04 23.7305C139.04 23.0818 139.581 22.5842 141.55 22.5842V22.8104ZM143.444 20.9099C143.444 19.2508 142.722 18.2705 140.423 18.2705C139.506 18.2705 138.469 18.4967 137.642 18.7983L137.928 20.2161C138.649 19.9294 139.521 19.7184 140.197 19.7184C141.204 19.7184 141.55 20.1407 141.55 21.0305V21.3472C138.935 21.3472 137.221 22.1918 137.221 23.9566C137.221 25.1782 138.018 26.0831 139.355 26.0831C140.483 26.0831 141.4 25.4798 141.76 24.6353L141.775 24.6502C141.715 25.0425 141.7 25.4949 141.7 25.9474C142.662 25.9775 143.804 25.9323 144.751 25.8116V24.6502C144.315 24.6202 143.804 24.5599 143.444 24.4994V20.9099ZM99.7185 21.3019C99.8235 20.2764 100.425 19.673 101.236 19.673C102.063 19.673 102.589 20.231 102.559 21.3019H99.7185ZM101.327 18.2856C99.1171 18.2856 97.8094 20.0048 97.8094 22.2523C97.8094 24.8917 99.1171 26.0831 101.492 26.0831C102.379 26.0831 103.326 25.9174 104.167 25.6307L103.972 24.2279C103.205 24.4845 102.454 24.6202 101.717 24.6202C100.32 24.6202 99.7184 23.9114 99.7032 22.5688H104.363C104.363 22.5688 104.423 22.0261 104.423 21.468C104.423 19.6883 103.506 18.2856 101.327 18.2856ZM91.3024 24.3485C90.9267 24.3485 90.5057 24.3336 90.2202 24.2883V17.3352C90.5358 17.3052 90.9867 17.2899 91.4375 17.2899C93.7824 17.2899 94.6691 18.6023 94.6691 20.8193C94.6691 23.2626 93.2712 24.3485 91.3024 24.3485ZM91.5729 15.6914C90.3254 15.6914 89.078 15.7517 88.2663 15.7517H87.0038V16.9281C87.3946 17.0187 87.8605 17.0941 88.3114 17.1243V24.5599C87.8605 24.5901 87.3946 24.6655 87.0038 24.7558V25.9323H88.3114C88.8073 25.9323 90.0698 25.9928 91.122 25.9928C94.4285 25.9928 96.7133 24.1676 96.7133 20.6686C96.7133 17.1091 94.7143 15.6914 91.5729 15.6914ZM110.705 18.2856C109.473 18.2856 108.601 18.9041 108.15 19.8089L108.135 19.7939C108.21 19.3867 108.24 18.7832 108.24 18.4213C107.248 18.3911 106.106 18.4364 105.159 18.557V19.7184C105.595 19.7485 106.106 19.8089 106.467 19.8693V25.9323H108.345V22.3125C108.345 20.9702 109.112 19.854 110.104 19.854C110.84 19.854 111.066 20.3368 111.066 21.1361V25.9323H112.93V20.6686C112.93 19.3565 112.388 18.2856 110.705 18.2856Z"
-        fill="#155429"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M58.0044 40.7322H45.0341V47.9998H58.0044C67.4367 47.9998 70.9141 47.2803 70.9141 47.2803V39.9282C70.9141 39.9282 67.4035 40.7322 58.0044 40.7322Z"
-        fill="#1D6B35"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M28.2166 38.8932L30.5744 44.4397C32.3222 44.4203 34.0973 44.4102 35.8965 44.4102C38.6269 44.4102 41.303 44.4332 43.9092 44.4771V39.4079L54.79 37.9947C54.79 37.9947 53.2782 37.7322 49.168 37.4305C45.0582 37.1288 39.1552 37.2388 39.1552 37.2388C39.1552 37.2388 36.7875 38.0375 35.8742 38.1974C34.9607 38.0375 32.5933 37.2388 32.5933 37.2388C32.5933 37.2388 26.7349 37.129 22.6249 37.4305C18.5206 37.7313 15.7869 38.1651 15.7707 38.1675L28.2166 38.8932Z"
-        fill="#1D6B35"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M27.689 39.5922L27.7093 44.4801C28.655 44.4637 29.6103 44.4502 30.5743 44.4397L28.2166 38.8932L15.7707 38.1675V39.5667C24.1735 39.5462 27.689 39.5922 27.689 39.5922Z"
-        fill="#155429"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M26.564 40.7322H13.7434C4.34446 40.7322 0.833496 39.9282 0.833496 39.9282V47.2803C0.833496 47.2803 4.31124 47.9998 13.7434 47.9998H26.564V40.7322Z"
-        fill="#1D6B35"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M1.98302 40.1214C1.19158 40.0101 0.83374 39.9282 0.83374 39.9282V47.2803C0.83374 47.2803 2.08732 47.5395 5.12832 47.7467C3.11096 45.4611 2.13647 42.6289 1.98302 40.1214Z"
-        fill="#155429"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M5.13794 42.4888H6.51626V42.9216H6.14279L7.02139 45.3659C7.05807 45.4661 7.07926 45.5823 7.10556 45.6877H7.11608C7.14223 45.5823 7.16342 45.4661 7.20024 45.3659L8.06292 42.9216H7.63684V42.4888H8.9836V42.9216H8.65221L7.38961 46.184H6.73719L5.45881 42.9216H5.13794V42.4888Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M10.7589 44.1469C11.1744 44.1469 11.4637 43.9252 11.4637 43.5293C11.4637 43.144 11.227 42.9222 10.801 42.9222H10.3118L10.3012 44.1469H10.7589ZM9.33325 45.7518H9.69621L9.71725 42.9221H9.33325V42.4893H10.9693C11.7269 42.4893 12.0846 42.9275 12.0846 43.4659C12.0846 44.1046 11.6743 44.3951 11.2062 44.4426V44.453C11.4112 44.469 11.5479 44.643 11.6743 44.8542L12.0741 45.53C12.1793 45.7042 12.2793 45.7518 12.4003 45.7518H12.5372V46.1847H12.1951C11.8848 46.1847 11.7427 46.153 11.5271 45.7939L11.0219 44.9335C10.8694 44.6696 10.7905 44.6114 10.6169 44.6114H10.3013L10.2908 45.7518H10.7168V46.1847H9.33325V45.7518Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M12.9893 45.7512H13.3575L13.3785 42.9216H12.9525V42.4888H15.5408L15.5618 43.46H15.0725L15.0409 42.9216H13.9729L13.9626 44.0619H15.0146V44.5264H13.9626L13.9519 45.7512H15.0725L15.1145 45.1493H15.6092L15.5671 46.184H12.9893V45.7512Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M17.7552 45.7517C18.6547 45.7517 19.1598 45.2026 19.1598 44.2787C19.1598 43.45 18.6863 42.9222 17.7815 42.9222H17.187L17.1661 45.7517H17.7552ZM16.2032 45.7517H16.5767L16.5977 42.9221H16.2032V42.4893H17.9497C19.1334 42.4893 19.812 43.1807 19.812 44.2524C19.812 45.3452 19.1229 46.1845 17.9445 46.1845H16.2032V45.7517Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M20.2274 45.7512H20.5956L20.6166 42.9216H20.1906V42.4888H22.7789L22.7999 43.46H22.3106L22.279 42.9216H21.2112L21.2007 44.0619H22.2529V44.5264H21.2007L21.19 45.7512H22.3106L22.3528 45.1493H22.8473L22.805 46.184H20.2274V45.7512Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M50.2324 44.1469C50.6479 44.1469 50.9372 43.9252 50.9372 43.5293C50.9372 43.144 50.7005 42.9222 50.2743 42.9222H49.7851L49.7746 44.1469H50.2324ZM48.8066 45.7518H49.1695L49.1906 42.9221H48.8066V42.4893H50.4425C51.2001 42.4893 51.5578 42.9275 51.5578 43.4659C51.5578 44.1046 51.1475 44.3951 50.6794 44.4426V44.453C50.8845 44.469 51.0213 44.643 51.1475 44.8542L51.5475 45.53C51.6525 45.7042 51.7524 45.7518 51.8736 45.7518H52.0103V46.1847H51.6684C51.3579 46.1847 51.2159 46.153 51.0002 45.7939L50.4953 44.9335C50.3426 44.6696 50.2638 44.6114 50.0902 44.6114H49.7745L49.7639 45.7518H50.1902V46.1847H48.8066V45.7518Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M52.4622 45.7512H52.8305L52.8515 42.9216H52.4254V42.4888H55.0137L55.0346 43.46H54.5454L54.5139 42.9216H53.446L53.4355 44.0619H54.4877V44.5264H53.4355L53.4249 45.7512H54.5454L54.5877 45.1493H55.0821L55.0399 46.184H52.4622V45.7512Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M58.1171 43.5709V42.9797C57.9803 42.9426 57.7698 42.8898 57.4911 42.8898C56.6125 42.8898 56.1547 43.4284 56.1547 44.3521C56.1547 45.1917 56.6599 45.7829 57.4122 45.7829C57.6646 45.7829 57.9328 45.7563 58.0698 45.7299L58.0908 45.0648H58.6168L58.5959 46.0732C58.3591 46.1524 57.8856 46.2474 57.449 46.2474C56.1286 46.2474 55.5026 45.4238 55.5026 44.3732C55.5026 43.2223 56.2706 42.4253 57.4962 42.4253C57.9593 42.4253 58.3959 42.5256 58.6431 42.6259V43.5709H58.1171Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M61.5152 45.7512H61.8835L61.8941 44.5634H60.1844L60.1737 45.7512H60.5683V46.184H59.2111V45.7512H59.5793L59.6004 42.9216H59.2111V42.4888H60.5683V42.9216H60.1948L60.1844 44.0987H61.8941L61.9046 42.9216H61.5152V42.4888H62.8725V42.9216H62.4991L62.478 45.7512H62.8725V46.184H61.5152V45.7512Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M64.2677 45.7512H64.6466L64.6674 42.9216H63.8099L63.7888 43.46H63.2628L63.2839 42.4888H66.6245L66.6454 43.46H66.1194L66.0984 42.9216H65.2567L65.2355 45.7512H65.6616V46.184H64.2677V45.7512Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M28.3715 5.93919C28.5313 5.83271 28.7492 5.75866 28.9844 5.7291C29.0546 5.72096 29.1249 5.71643 29.1945 5.71643C29.6391 5.71643 29.9596 5.89364 29.9925 6.15726C30.0304 6.46809 29.6538 6.75177 29.1174 6.81768C29.0463 6.82643 28.9761 6.8308 28.9068 6.8308C28.4621 6.8308 28.1412 6.65344 28.1092 6.38937C28.0891 6.22528 28.1821 6.06572 28.3715 5.93919ZM32.426 5.35831L33.5332 5.89325L32.4555 6.48656L31.348 5.95102L32.426 5.35831ZM35.8307 5.2552C36.4122 5.2552 36.8509 5.50646 36.8509 5.83976C36.8509 6.17305 36.4122 6.42446 35.8307 6.42446C35.2489 6.42446 34.8102 6.17305 34.8102 5.83976C34.8102 5.50646 35.2489 5.2552 35.8307 5.2552ZM39.2357 5.35831L40.3133 5.95102L39.2064 6.48656L38.1281 5.89371L39.2357 5.35831ZM41.6691 6.15772C41.7014 5.89364 42.0223 5.71643 42.467 5.71643C42.5363 5.71643 42.6072 5.72096 42.6777 5.7291C42.9129 5.75866 43.1298 5.83271 43.2899 5.93919C43.4789 6.06572 43.5724 6.22528 43.5521 6.38937C43.5198 6.65314 43.2 6.8308 42.7554 6.8308C42.686 6.8308 42.6147 6.82643 42.5444 6.81768C42.0073 6.75177 41.6315 6.46809 41.6691 6.15772ZM36.0602 7.6007C39.5777 7.6007 42.229 7.77338 43.9926 7.94501L44.3928 4.7298C42.6688 4.57687 39.4251 4.19531 35.8988 4.19531C32.3396 4.19531 28.9971 4.5817 27.2412 4.73402L27.6589 7.99282C29.4062 7.80701 32.2147 7.6007 36.0602 7.6007Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M36.8644 1.90625C36.3935 1.90625 36.0121 2.28947 36.0121 2.76152C36.0121 3.13207 36.2465 3.44727 36.5751 3.56642L37.4137 2.10819C37.2652 1.98256 37.0737 1.90625 36.8644 1.90625ZM27.0798 2.48477C26.609 2.48477 26.2281 2.86769 26.2281 3.34004C26.2281 3.81209 26.609 4.19531 27.0798 4.19531C27.5508 4.19531 27.9323 3.81209 27.9323 3.34004C27.9323 2.86769 27.5508 2.48477 27.0798 2.48477ZM29.0374 2.34008C28.5665 2.34008 28.1848 2.7233 28.1848 3.19535C28.1848 3.6674 28.5665 4.05062 29.0374 4.05062C29.5082 4.05062 29.8894 3.6674 29.8894 3.19535C29.8894 2.7233 29.5082 2.34008 29.0374 2.34008ZM30.9936 2.19563C30.5235 2.19563 30.1414 2.57839 30.1414 3.05074C30.1414 3.52294 30.5235 3.90616 30.9936 3.90616C31.465 3.90616 31.8465 3.52294 31.8465 3.05074C31.8465 2.57839 31.465 2.19563 30.9936 2.19563ZM32.9506 2.05117C32.4798 2.05117 32.0983 2.43394 32.0983 2.90629C32.0983 3.37879 32.4798 3.76125 32.9506 3.76125C33.4215 3.76125 33.8028 3.37879 33.8028 2.90629C33.8028 2.43394 33.4215 2.05117 32.9506 2.05117ZM34.907 1.90625C34.4372 1.90625 34.0547 2.28947 34.0547 2.76152C34.0547 3.23402 34.4372 3.61679 34.907 3.61679C35.3777 3.61679 35.7596 3.23402 35.7596 2.76152C35.7596 2.28947 35.3777 1.90625 34.907 1.90625ZM27.6441 1.43409C27.6441 0.961737 27.2626 0.578518 26.7916 0.578518C26.3207 0.578518 25.9395 0.961737 25.9395 1.43409C25.9395 1.90614 26.3207 2.28905 26.7916 2.28905C27.2626 2.28905 27.6441 1.90614 27.6441 1.43409ZM36.738 0.855268C36.738 0.383219 36.3566 0 35.8857 0C35.415 0 35.0334 0.383219 35.0334 0.855268C35.0334 1.32822 35.415 1.71054 35.8857 1.71054C36.3566 1.71054 36.738 1.32822 36.738 0.855268Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M37.4135 2.10817L36.5749 3.56639C36.6651 3.59927 36.7626 3.61676 36.8642 3.61676C37.3351 3.61676 37.7167 3.234 37.7167 2.76149C37.7167 2.49953 37.5988 2.26501 37.4135 2.10817ZM38.8207 2.05127C38.3504 2.05127 37.9683 2.43403 37.9683 2.90638C37.9683 3.37889 38.3504 3.76135 38.8207 3.76135C39.2921 3.76135 39.6736 3.37889 39.6736 2.90638C39.6736 2.43403 39.2921 2.05127 38.8207 2.05127ZM40.7775 2.19572C40.3067 2.19572 39.9252 2.57849 39.9252 3.05084C39.9252 3.52304 40.3067 3.90626 40.7775 3.90626C41.2489 3.90626 41.63 3.52304 41.63 3.05084C41.63 2.57849 41.2489 2.19572 40.7775 2.19572ZM42.735 2.34018C42.2641 2.34018 41.8828 2.7234 41.8828 3.19545C41.8828 3.66749 42.2641 4.05071 42.735 4.05071C43.2054 4.05071 43.5868 3.66749 43.5868 3.19545C43.5868 2.7234 43.2054 2.34018 42.735 2.34018ZM44.6915 2.48487C44.221 2.48487 43.8392 2.86778 43.8392 3.34013C43.8392 3.81218 44.221 4.1954 44.6915 4.1954C45.1628 4.1954 45.5435 3.81218 45.5435 3.34013C45.5435 2.86778 45.1628 2.48487 44.6915 2.48487ZM44.9795 0.578613C44.5089 0.578613 44.1272 0.961832 44.1272 1.43418C44.1272 1.90623 44.5089 2.28915 44.9795 2.28915C45.4504 2.28915 45.8321 1.90623 45.8321 1.43418C45.8321 0.961832 45.4504 0.578613 44.9795 0.578613Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M39.2362 5.3582L40.3137 5.9509L39.2069 6.48629L38.1285 5.89359L39.2362 5.3582ZM41.6696 6.15784C41.7019 5.89376 42.0227 5.71655 42.4675 5.71655C42.5367 5.71655 42.6077 5.72093 42.6782 5.72922C42.9134 5.75863 43.1302 5.83283 43.2903 5.93931C43.4794 6.06569 43.5727 6.2254 43.5526 6.38949C43.5202 6.65326 43.2004 6.83092 42.7559 6.83092C42.6864 6.83092 42.6152 6.8264 42.5449 6.8178C42.0077 6.75174 41.632 6.46806 41.6696 6.15784ZM35.8312 6.42427C35.4865 6.42427 35.1919 6.33574 35.009 6.19458L34.1664 7.61766C34.7676 7.60665 35.3987 7.60047 36.0606 7.60047C39.5781 7.60047 42.2294 7.7733 43.9928 7.94478L44.393 4.72972C42.7165 4.58117 39.6037 4.21605 36.1915 4.19629L35.5523 5.27552C35.6405 5.2624 35.7335 5.25501 35.8312 5.25501C36.4127 5.25501 36.8514 5.50626 36.8514 5.83956C36.8514 6.17286 36.4127 6.42427 35.8312 6.42427Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M43.9091 39.4079V39.5922H56.0221V38.1673C56.0221 38.1673 55.5779 38.0975 54.7899 37.9946L43.9091 39.4079Z"
-        fill="#155429"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M70.1289 47.406C70.6673 47.3225 70.9142 47.2803 70.9142 47.2803V39.9282C70.9142 39.9282 70.0488 40.1262 68.0284 40.3258C69.4005 42.0027 69.9481 44.7336 70.1289 47.406Z"
-        fill="#155429"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M32.8755 42.1995H33.1941L33.2123 39.7506H32.8436V39.376H35.0836L35.1018 40.2168H34.6784L34.651 39.7506H33.7269L33.7177 40.7375H34.6283V41.1396H33.7177L33.7087 42.1995H34.6784L34.7148 41.6788H35.1428L35.1064 42.5741H32.8755V42.1995Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M35.6751 42.1995H35.9847L36.0029 39.7506H35.6022V39.376H36.4399L37.915 41.3087C38.0107 41.4366 38.1472 41.6466 38.2338 41.7884H38.2429L38.2248 39.7506H37.8696V39.376H38.9804V39.7506H38.671L38.6526 42.5741H38.2384L36.6447 40.4998C36.5538 40.3812 36.4673 40.2533 36.4217 40.1619L36.4125 40.1711L36.4307 42.1995H36.7814V42.5741H35.6751V42.1995Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M58.926 33.2282H58.9268L58.926 33.2246V33.2282Z"
-        fill="url(#paint0_linear_10499_33040)"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M70.0351 17.7954C70.0351 16.144 67.8189 16.6658 66.6035 16.6658C65.3876 16.6658 65.4743 15.104 65.4743 15.104H67.7883C69.6351 15.104 70.8162 16.2364 70.0351 17.7954Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M62.8545 14.4162C63.3944 13.879 63.9546 13.583 64.7512 13.583C66.5195 13.583 69.077 13.471 69.4339 12.5474C69.7443 12.8711 69.7155 13.5248 68.9598 14.276C70.9508 14.276 71.3013 14.7264 71.5022 15.1475C72.0218 16.236 71.1203 17.7952 70.0349 17.7952C70.816 16.2365 69.635 15.1039 67.7881 15.1039H65.4742H65.1279H64.7512C64.4801 15.1039 64.1829 15.2452 63.9356 15.4912C63.4112 16.0129 63.4112 16.862 63.9356 17.3836L67.6014 21.0298C67.8329 21.2599 68.0001 21.5212 68.1362 21.7995L62.1057 17.1995C61.84 16.2401 62.0977 15.1693 62.8545 14.4162Z"
-        fill="#EFD544"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M58.5332 10.8457C58.5038 10.7362 58.485 10.6332 58.4747 10.5355C59.0067 9.91477 59.8945 9.67055 61.0939 10.1741C62.6034 10.8075 63.979 9.71045 63.979 9.71045C62.9409 11.6887 61.4995 11.0469 60.2495 10.6563C59.5669 10.4433 58.9263 10.6527 58.5332 10.8457Z"
-        fill="#B98F16"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M52.1796 18.2863C50.8966 17.5568 49.8652 17.025 49.8631 17.0241L49.8628 17.0239L50.351 19.1245C50.351 19.1245 48.9027 21.4686 48.1269 22.7407C48.1435 22.749 48.161 22.7554 48.1776 22.7643C48.1776 22.7643 52.1868 18.2906 52.1867 18.2905V18.2903C52.1842 18.289 52.182 18.2876 52.1796 18.2863Z"
-        fill="#B98F16"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M59.1167 14.9694C59.1167 13.4313 58.0731 12.8007 57.6063 12.595C56.9143 12.5668 56.2008 12.6312 55.5887 12.8796C55.7885 14.6982 58.4148 17.129 59.9929 17.6659C60.0056 17.6696 60.0173 17.6737 60.0302 17.6776C59.2764 17.0993 60.5298 14.9694 59.1167 14.9694Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M53.781 19.2298C53.2424 18.8994 52.6953 18.5796 52.1869 18.2905L52.1867 18.2907L48.1776 22.7643C47.9552 22.6464 47.7063 22.5726 47.437 22.5726V24.1404H47.4394C46.7632 24.2888 46.3115 24.8459 46.3115 25.7055H47.4402L47.4394 27.1039C47.5887 27.2692 47.8238 27.3747 48.0629 27.3816L53.7876 19.2338C53.7851 19.2328 53.7795 19.2317 53.781 19.2298Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M59.6778 19.1758C58.1159 20.1506 55.0021 20.0113 54.9991 20.0094C54.9975 20.0082 54.9958 20.0071 54.9942 20.0061C54.9905 20.0074 55.8734 20.5446 56.7864 21.42L62.9184 22.3911C62.9184 22.3911 60.9388 20.4269 59.6778 19.1758Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M66.374 30.157C65.6651 30.157 63.5231 29.8757 63.5231 28.312V25.0728L61.8876 27.7256L61.8662 30.4123C64.5077 33.4266 63.8459 36.8635 62.6409 38.2482V39.2882C63.3905 39.2738 64.1271 39.2564 64.833 39.2307L64.6489 38.4492L65.6274 36.7606C65.6274 36.7606 65.2 31.3245 66.374 30.157Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M53.7876 19.2339L48.0628 27.3815C48.0706 27.3815 48.0782 27.3828 48.0869 27.3828C48.1829 27.3828 50.1119 25.9571 50.1119 25.9571L54.9989 20.0093C54.6155 19.751 54.2043 19.4895 53.7876 19.2339Z"
-        fill="#EFD544"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M55.5887 12.8794C53.4727 13.9674 52.5031 17.1121 49.8629 17.0241C49.8667 17.0257 52.8635 18.5707 54.9989 20.009C55.003 20.0119 60.5499 20.2589 60.5499 17.7749C59.1419 17.7749 55.8143 14.935 55.5887 12.8794Z"
-        fill="#F6E477"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M50.2644 15.152C50.723 13.972 50.8912 12.0526 51.3937 10.147L48.9933 8.62875L48.4171 7.63477H45.7217L45.2204 8.49839C46.2612 8.59814 46.9272 8.67345 46.9477 8.6754L47.4184 8.73465V12.4787L50.2644 15.152Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M53.5169 6.03197V4.42474L55.1788 5.15245C55.164 5.06724 55.1289 4.61691 55.9111 4.18457C55.877 4.66911 56.5869 5.02539 57.4068 5.27456C56.5566 7.23987 55.2999 10.2479 55.5888 12.8794C53.4732 13.9675 52.5033 17.1122 49.863 17.0241C49.863 17.0241 49.1887 16.7056 48.9633 16.3153C51.4671 16.3153 50.1104 9.42089 53.5169 6.03197Z"
-        fill="#EFD544"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M60.3877 32.0976C61.2023 32.9106 61.6911 33.8525 61.6911 34.7622C61.6911 36.8143 60.2941 37.5609 60.2941 37.5609C58.6369 37.0326 58.1723 38.7064 58.0572 39.3143C58.6431 39.3152 59.2531 39.3155 59.8878 39.3155C60.7948 39.3155 61.7814 39.3059 62.6409 39.2903V38.2478C63.846 36.8634 64.5077 33.4263 61.8663 30.4121L60.3877 32.0976Z"
-        fill="#EFD544"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M58.7697 29.2327C58.7693 28.9004 58.7689 28.3391 58.7685 28.0195L58.7682 28.0197C58.7683 28.0963 58.7683 28.1731 58.7685 28.2511L53.6321 29.4403L50.2315 36.1008C51.2516 36.1833 52.1675 36.2717 52.9614 36.3569L52.6341 35.4948L56.7799 29.6113L58.7697 29.2327Z"
-        fill="#EFD544"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M51.2341 29.9956C50.8515 33.3221 48.6572 34.3769 48.6572 34.3769C48.6572 34.3769 48.2383 34.2735 47.8043 34.2735C46.8392 34.2735 46.3672 35.8608 46.3672 35.8608C46.8765 35.8836 47.4101 35.914 47.944 35.9471C48.397 35.9711 48.8401 35.9985 49.2639 36.029C49.5967 36.0534 49.9164 36.0783 50.23 36.1044L53.6324 29.4404L51.2341 29.9956Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M58.7663 26.1761L56.5581 24.5659L50.3994 29.6374C50.3994 29.6374 50.7196 30.0813 51.2342 29.9955L58.7686 28.2511C58.7676 27.4591 58.7669 26.73 58.7663 26.1761Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M54.9989 20.0093L53.2421 22.1475L58.7662 26.1761C58.7656 25.5404 58.7652 25.1356 58.7652 25.1356C58.7652 24.789 58.7062 24.4495 58.6034 24.1208C58.2823 23.0614 57.5255 22.1285 56.7865 21.4197C55.9023 20.5721 55.049 20.0432 54.9991 20.0093H54.9989Z"
-        fill="#EFD544"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M58.3082 6.8985C57.9364 6.91111 57.626 6.67154 57.6162 6.36386C57.615 6.32921 57.6191 6.29516 57.625 6.2617C58.083 6.29006 58.557 6.35156 58.9525 6.41397C58.9068 6.67769 58.6398 6.8871 58.3082 6.8985ZM61.4428 7.80981L60.0199 6.75522L59.7786 5.74804C59.7786 5.74804 58.4951 5.60598 57.4065 5.2749C56.5566 7.24006 55.2998 10.2483 55.5888 12.8798C57.5457 12.086 60.5377 13.1681 60.5377 13.1681L60.0678 11.9093L60.7546 11.1185C60.7546 11.1185 58.7161 11.3547 58.7153 11.3535C58.1136 10.0073 58.6171 9.34428 59.9283 9.34428C60.4753 9.34428 60.8997 9.608 60.8997 9.608L61.1835 8.68933L60.938 8.23419L61.2847 8.26599L61.4428 7.80981Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M12.9673 33.2228H12.9679V33.2192L12.9673 33.2228Z"
-        fill="url(#paint1_linear_10499_33040)"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M1.61038 17.792C1.61038 16.1406 3.81566 16.6624 5.02507 16.6624C6.23494 16.6624 6.14865 15.1006 6.14865 15.1006H3.84612C2.00836 15.1006 0.833165 16.233 1.61038 17.792Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M8.75539 14.4128C8.21813 13.8756 7.66078 13.5796 6.8681 13.5796C5.10853 13.5796 2.56364 13.4676 2.20843 12.5439C1.89958 12.8677 1.92825 13.5214 2.68025 14.2725C0.699023 14.2725 0.35026 14.723 0.150367 15.1441C-0.366625 16.2326 0.530344 17.7918 1.6104 17.7918C0.833186 16.233 2.00838 15.1005 3.84614 15.1005H6.14867H6.49323H6.8681C7.13778 15.1005 7.43357 15.2418 7.67968 15.4878C8.20148 16.0095 8.20148 16.8586 7.67968 17.3802L4.03193 21.0264C3.80157 21.2565 3.63514 21.5178 3.49978 21.7961L9.50048 17.1961C9.76491 16.2366 9.50844 15.1659 8.75539 14.4128Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M13.0555 10.8423C13.0847 10.7328 13.1035 10.6298 13.1137 10.5321C12.5844 9.91135 11.7009 9.66713 10.5074 10.1707C9.00539 10.8041 7.6366 9.70703 7.6366 9.70703C8.66953 11.6852 10.1039 11.0435 11.3477 10.6529C12.0269 10.4398 12.6644 10.6493 13.0555 10.8423Z"
-        fill="#B98F16"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M19.3778 18.2829C20.6544 17.5534 21.6808 17.0216 21.6829 17.0207L21.6832 17.0205L21.1974 19.1211C21.1974 19.1211 22.6385 21.4652 23.4105 22.7373C23.394 22.7456 23.3765 22.752 23.36 22.7609L19.3707 18.2871V18.2869C19.3731 18.2856 19.3754 18.2842 19.3778 18.2829Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M13.9775 12.5916C13.5131 12.7972 12.4746 13.4279 12.4746 14.9659C11.0684 14.9659 12.3158 17.0958 11.5659 17.6741C11.5785 17.6702 11.5902 17.6662 11.6028 17.6624C13.1732 17.1255 15.7866 14.6947 15.9852 12.8761C15.376 12.6277 14.6661 12.5635 13.9775 12.5916Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M17.7841 19.2263C18.32 18.896 18.8645 18.5762 19.3704 18.2871L19.3705 18.2873L23.3598 22.7609C23.5812 22.643 23.8288 22.5692 24.0968 22.5692V24.137H24.0944C24.7673 24.2853 25.2168 24.8425 25.2168 25.7021H24.0937L24.0944 27.1005C23.9459 27.2658 23.7119 27.3712 23.474 27.3781L17.7775 19.2304C17.7801 19.2293 17.7856 19.2283 17.7841 19.2263Z"
-        fill="#EFD544"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M11.9165 19.1724C13.4708 20.1471 16.5691 20.0079 16.5721 20.006C16.5738 20.0048 16.5754 20.0037 16.5771 20.0027C16.5807 20.004 15.7022 20.5412 14.7937 21.4165L8.69196 22.3877C8.69196 22.3877 10.6618 20.4235 11.9165 19.1724Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M5.25311 30.1531C5.95859 30.1531 8.09004 29.8718 8.09004 28.3081V25.0688L9.7174 27.7217L9.73871 30.4084C7.11023 33.4227 7.76874 36.8596 8.9678 38.2443V39.2843C8.22195 39.2699 7.48901 39.2525 6.78653 39.2268L6.96976 38.4453L5.99611 36.7567C5.99611 36.7567 6.42141 31.3206 5.25311 30.1531Z"
-        fill="#F6E477"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M8.96785 39.287C9.8231 39.3026 10.8049 39.3122 11.7075 39.3122C12.339 39.3122 12.9461 39.3119 13.5291 39.311C13.4146 38.703 12.9522 37.0293 11.3032 37.5577C11.3032 37.5577 9.91314 36.8111 9.91314 34.7587C9.91314 33.8465 10.3946 32.908 11.2094 32.0935L9.73876 30.4087C7.11028 33.423 7.76879 36.86 8.96785 38.2446V39.287Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M17.7776 19.23L23.4741 27.3776C23.4663 27.3776 23.4589 27.3789 23.4501 27.3789C23.3546 27.3789 21.4351 25.9532 21.4351 25.9532L16.5722 20.0054C16.9537 19.7471 17.3629 19.4856 17.7776 19.23Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M11.0488 17.7715C11.0488 20.2557 16.5682 20.0085 16.5724 20.0057C18.6972 18.5674 21.6793 17.0223 21.6829 17.0207C19.0558 17.1087 18.0911 13.964 15.9855 12.876C15.7608 14.9316 12.4498 17.7715 11.0488 17.7715Z"
-        fill="#F6E477"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M21.2834 15.1486C20.827 13.9686 20.6597 12.0492 20.1597 10.1436L22.5482 8.62533L23.1216 7.63135H25.8038L26.3025 8.49497C25.2669 8.59473 24.6041 8.67003 24.5837 8.67198L24.1154 8.73124V12.4752L21.2834 15.1486Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M18.0469 6.02855V4.42132L16.3933 5.14903C16.408 5.06382 16.4431 4.61349 15.6646 4.18115C15.6985 4.66569 14.9921 5.02197 14.1763 5.27114C15.0223 7.23645 16.2728 10.2445 15.9853 12.876C18.0905 13.964 19.0556 17.1087 21.6828 17.0207C21.6828 17.0207 22.3538 16.7022 22.5782 16.3119C20.0867 16.3119 21.4367 9.41747 18.0469 6.02855Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M4.04572 25.0691C4.58537 25.6091 5.30376 25.9063 6.06792 25.9063C6.83162 25.9063 7.55031 25.6091 8.09011 25.0691L9.71747 27.7219L9.73878 30.4086L12.8146 33.9322C12.8161 32.9509 12.8247 25.1322 12.8247 25.1322C12.8247 24.7855 12.8833 24.4461 12.9855 24.1174C13.305 23.058 14.0582 22.1251 14.7936 21.4163L8.69189 22.3875L7.44166 23.6161C7.44166 23.6161 6.09793 24.9705 5.12202 23.9941C4.59963 23.472 4.59963 22.6233 5.12202 22.1017L8.76918 18.4558C9.13145 18.0938 9.37366 17.6577 9.50047 17.1963L3.49977 21.7963C2.98157 22.8624 3.16106 24.1846 4.04572 25.0691ZM14.7943 21.4156C14.7942 21.4159 14.7939 21.4161 14.7937 21.4162C14.7939 21.4161 14.7942 21.4159 14.7943 21.4156Z"
-        fill="#EFD544"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M12.8201 29.2293C12.8205 28.897 12.8208 28.3356 12.8213 28.0161L12.8216 28.0163C12.8214 28.0929 12.8214 28.1697 12.8213 28.2477L17.9324 29.4369L21.3161 36.0974C20.3011 36.1799 19.3897 36.2683 18.5997 36.3535L18.9254 35.4914L14.8001 29.6079L12.8201 29.2293Z"
-        fill="#F6E477"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M20.3186 29.9922C20.6993 33.3187 22.8829 34.3735 22.8829 34.3735C22.8829 34.3735 23.2996 34.2701 23.7315 34.2701C24.6918 34.2701 25.1615 35.8574 25.1615 35.8574C24.6547 35.8802 24.1238 35.9106 23.5925 35.9436C23.1417 35.9676 22.7008 35.9951 22.2791 36.0255C21.9479 36.05 21.6298 36.0749 21.3178 36.101L17.9322 29.437L20.3186 29.9922Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M12.8235 26.1727L15.0208 24.5625L21.1491 29.634C21.1491 29.634 20.8305 30.0778 20.3184 29.992L12.8212 28.2477C12.8221 27.4556 12.8229 26.7266 12.8235 26.1727Z"
-        fill="#E2C029"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M16.5723 20.0059L18.3205 22.1441L12.8235 26.1727C12.8241 25.537 12.8246 25.1322 12.8246 25.1322C12.8246 24.7855 12.8833 24.4461 12.9856 24.1174C13.3051 23.058 14.0582 22.1251 14.7935 21.4163C15.6734 20.5687 16.5225 20.0398 16.5721 20.0059H16.5723Z"
-        fill="#B98F16"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M13.2794 6.895C12.9493 6.88345 12.6837 6.67419 12.6383 6.41032C13.0317 6.34806 13.5034 6.28656 13.9592 6.25805C13.965 6.29166 13.9691 6.32571 13.9679 6.36036C13.9581 6.66804 13.6493 6.90746 13.2794 6.895ZM14.1765 5.27148C13.0933 5.60241 11.8162 5.74462 11.8162 5.74462L11.5761 6.75165L10.1602 7.80639L10.3175 8.26257L10.6625 8.23077L10.4182 8.68591L10.7006 9.60458C10.7006 9.60458 11.1229 9.34086 11.6672 9.34086C12.9719 9.34086 13.473 10.0038 12.8742 11.3501C12.8735 11.3513 10.845 11.1151 10.845 11.1151L11.5286 11.9059L11.0608 13.1647C11.0608 13.1647 14.038 12.0826 15.9853 12.8763C16.2729 10.2448 15.0223 7.23664 14.1765 5.27148Z"
-        fill="#EFD544"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M46.2697 28.7936C46.2697 34.6975 35.7429 36.7886 35.7429 36.7886C35.7429 36.7886 25.217 34.6975 25.217 28.7936V9.74011C25.217 9.74011 28.7789 9.06445 35.7429 9.06445C42.708 9.06445 46.2697 9.74011 46.2697 9.74011V28.7936Z"
-        fill="#D2AD15"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M35.7429 36.7885C35.7429 36.7885 46.2697 34.6975 46.2697 28.7934V9.73996C46.2697 9.73996 42.7078 9.06445 35.7429 9.06445C34.1024 9.06445 32.6507 9.10196 31.3898 9.15926C40.8936 9.53879 44.6918 10.8232 44.6918 10.8232C44.6918 10.8232 44.7004 24.1138 44.7004 28.9841C44.7004 33.8544 35.7429 36.7885 35.7429 36.7885Z"
-        fill="#B98F16"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M37.8685 20.6003L37.0628 20.5001L36.8135 20.8325L34.9315 19.6333C35.1207 20.58 35.6371 22.3836 36.9613 23.9179C37.8434 24.5559 38.7654 25.224 39.4975 25.7539C39.4397 22.6566 38.3906 20.9513 38.1176 20.2679L37.8685 20.6003Z"
-        fill="#4B4B4B"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M34.0162 21.7923C33.0299 21.0886 33.3527 19.6965 34.0015 19.0415L36.8134 20.8327L37.0626 20.5002L37.8684 20.6004L38.1175 20.268L40.1968 20.6256C39.8086 20.0048 39.1621 19.0051 38.1798 18.3357L38.429 18.0034C37.9269 17.767 36.779 17.5795 35.4519 17.5795H33.4281L34.8583 13.5038C35.2387 12.4068 34.5611 11.5292 33.8237 11.4478C32.7094 11.3252 32.2485 12.5235 32.2485 12.5235L33.4706 13.7678C33.4982 13.7742 31.8947 17.5125 31.8947 17.5125C31.7309 17.904 31.6238 18.4113 31.6202 18.9595C31.6134 20.0063 32.092 21.0749 33.0672 22.0218C33.8944 22.8251 37.4743 25.6319 37.4743 25.6319L37.7624 25.2492L39.4974 25.7541C39.4974 25.7541 34.2917 21.9884 34.0162 21.7923Z"
-        fill="white"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M34.9313 19.6335L34.0016 19.0415C33.3527 19.6965 33.0299 21.0886 34.0161 21.7923C34.3163 22.0057 35.5889 22.9256 36.9611 23.9181C35.6369 22.3838 35.1206 20.5803 34.9313 19.6335Z"
-        fill="#636363"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M40.1968 20.6257L38.1176 20.2681C38.3904 20.9515 39.4396 22.6569 39.4975 25.7542C40.3275 26.3554 40.913 26.779 40.913 26.779L41.3234 26.2324L42.3991 26.9765C42.3403 26.4834 42.257 26.0191 42.1549 25.5792C41.6885 23.3928 40.917 21.7475 40.1968 20.6257Z"
-        fill="black"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M38.3749 33.4062C38.3289 33.2904 38.243 33.1797 38.0784 33.1797H37.4742V30.3218L36.9793 30.5466V33.1711H36.0615C35.9549 33.1711 35.7493 33.3223 35.6788 33.3223H35.2686C35.1033 33.3223 35.0171 33.4324 34.9717 33.5487L34.9416 33.6387L34.9245 33.7524C34.9208 33.7889 34.9208 33.8133 34.9208 33.8133H35.9103L36.1137 33.6624H36.5265C36.2665 33.822 35.9882 33.9935 35.8827 34.0577C35.7414 34.1442 35.7241 34.2842 35.7447 34.4071L35.7666 34.5025L35.809 34.6054C35.8249 34.639 35.8365 34.66 35.8365 34.66L37.5132 33.6702H38.4258C38.4258 33.6702 38.4252 33.6462 38.4218 33.6094L38.3934 33.4917L38.3749 33.4062Z"
-        fill="#4B4B4B"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M29.6797 22.8035L29.5866 22.9281L29.642 23.7398L29.7969 23.5309L29.9084 23.5444L30.5481 24.6542C30.5481 24.6542 30.5564 24.6489 30.5699 24.6407L30.6493 24.5757L30.7187 24.5139C30.8112 24.4137 30.8891 24.268 30.796 24.1053C30.771 24.0619 30.7298 23.9899 30.6791 23.9017L32.5705 24.7992L33.3436 24.6167L30.2444 23.1492C30.1758 23.0316 30.0817 22.8528 30.0817 22.8528C30.0265 22.7673 29.821 22.7354 29.6797 22.8035Z"
-        fill="#636363"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M33.4706 13.7678L32.2485 12.5234L26.9834 13.7678H33.4706Z"
-        fill="black"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M29.4856 23.064L28.4783 23.1931C28.4783 23.1931 28.1776 23.5785 28.0241 23.7846C27.9253 23.9176 27.9609 24.0533 28.0256 24.16L28.0865 24.2389L28.1585 24.3203C28.1848 24.3446 28.2042 24.3595 28.2042 24.3595L28.7174 23.6696L29.1316 23.6109C28.9333 23.8781 28.7219 24.1621 28.6482 24.262C28.5495 24.3949 28.5853 24.531 28.6506 24.6375L28.716 24.7216L28.7828 24.7975C28.8092 24.8224 28.8286 24.8373 28.8286 24.8373L29.6418 23.7399L29.5865 22.9282L29.4856 23.064Z"
-        fill="#4B4B4B"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M36.275 24.6831C36.1176 24.5578 35.9516 24.4252 35.7804 24.2881V25.7264L35.7072 25.7354L33.3436 24.6166L32.5703 24.7992L35.5178 26.1982L35.8202 26.5878L36.305 26.5182L36.488 26.0591L36.275 25.7852V24.6831Z"
-        fill="#4B4B4B"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M34.1618 12.5193C34.1618 12.5193 33.5983 12.7713 33.4603 12.7713C33.323 12.7713 33.1943 12.6583 33.1943 12.5193C33.1943 12.38 33.323 12.2671 33.4603 12.2671C33.5983 12.2671 34.1618 12.5193 34.1618 12.5193Z"
-        fill="black"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M28.3912 15.643C28.0378 15.0175 28.4515 14.6351 28.7421 14.3415C28.9522 14.128 29.1343 13.9449 29.1382 13.6754C29.1435 13.3217 28.9424 13.2571 28.4479 13.1854C28.1541 13.1426 27.8503 13.0985 27.6773 12.8757C27.5796 12.7484 27.5397 12.5999 27.5627 12.4467C27.614 12.1047 27.9648 11.8491 28.0043 11.8211C28.1035 11.7517 28.2387 11.7766 28.3069 11.8758C28.3757 11.9751 28.3516 12.1117 28.2531 12.1812C28.1754 12.2365 28.0103 12.3895 27.992 12.5127C27.9887 12.5373 27.9895 12.5657 28.0205 12.6059C28.0847 12.6896 28.3103 12.7221 28.5093 12.7512C28.9373 12.8133 29.5838 12.9072 29.5724 13.6817C29.5661 14.1289 29.2918 14.4066 29.0498 14.6516C28.7742 14.93 28.5495 15.1336 28.6985 15.5267C28.7962 15.7845 29.2383 15.9211 29.1534 16.0737C29.0733 16.2174 28.5644 15.9496 28.3912 15.643Z"
-        fill="black"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M37.4742 25.6321C37.4742 25.6321 37.2824 25.4818 36.9793 25.2422V26.1489H36.9851L36.7399 26.5771L36.9793 26.9967V30.5471L37.4742 30.3222V27.0053L37.7194 26.5771L37.4742 26.1489V25.6321Z"
-        fill="black"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M41.1624 26.4463L40.9131 26.7787L41.1624 26.4463Z"
-        fill="#636363"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M41.0719 19.83C40.3176 19.0498 39.4234 18.4218 38.4293 18.0034L38.1799 18.3358C39.1622 19.005 39.8087 20.0049 40.197 20.6256C40.9172 21.7474 41.6887 23.3928 42.1551 25.5793C42.257 26.019 42.3404 26.4833 42.3992 26.9764L43.4386 27.7048C43.9687 24.7001 42.9619 21.7848 41.0719 19.83Z"
-        fill="#636363"
-      ></path>
-      <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
-        d="M67.9874 21.7992L61.9868 17.1992C62.1135 17.6607 62.3559 18.0967 62.7181 18.4587L66.3651 22.1046C66.8875 22.6264 66.8875 23.475 66.3651 23.997C65.3894 24.9734 64.0455 23.6191 64.0455 23.6191L62.7953 22.3904L56.6936 21.4194C57.4291 22.128 58.1821 23.0609 58.5016 24.1205C58.604 24.449 58.6625 24.7886 58.6625 25.1352C58.6625 25.1352 58.671 32.9538 58.6727 33.9352L61.7485 30.4115L61.7697 27.7248L63.397 25.072C63.937 25.612 64.6555 25.9092 65.4194 25.9092C66.1834 25.9092 66.9019 25.612 67.4414 25.072C68.3261 24.1877 68.5056 22.8653 67.9874 21.7992ZM56.6935 21.4194C56.6933 21.4191 56.693 21.419 56.6929 21.4188C56.693 21.419 56.6933 21.4191 56.6935 21.4194Z"
-        fill="#E2C029"
-      ></path>
-      <defs>
-        <linearGradient
-          id="paint0_linear_10499_33040"
-          x1="58.9277"
-          y1="33.2279"
-          x2="58.9265"
-          y2="33.2279"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#B88E15"></stop>
-          <stop offset="1" stop-color="#D1AC14"></stop>
-        </linearGradient>
-        <linearGradient
-          id="paint1_linear_10499_33040"
-          x1="12.9671"
-          y1="33.2236"
-          x2="12.9683"
-          y2="33.2236"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#B88E15"></stop>
-          <stop offset="1" stop-color="#D1AC14"></stop>
-        </linearGradient>
-      </defs>
-    </svg>
-  );
-};
