+++ packages/storybook/src/templates/form-wmebv/wmebv-2-login.stories.tsx
@@ -13,9 +13,8 @@ import {
   Paragraph,
 } from '@utrecht/component-library-react/dist/css-module';
 import { ReactElement } from 'react';
-import { Layout } from './components/Layout';
-import { DenHaagLogo, PageHeaderLogo } from './components/Logo';
-import { VoorbeeldFooterLogo } from '../mijn-profiel/components/Logo';
+import { Layout } from '../../components/Layout';
+import { DenHaagLogo, PageHeaderLogo, VoorbeeldFooterLogo } from '../../components/Logo';
 
 const meta = {
   title: 'Templates/Meerstappenformulier/WMEBV Contactformulier/2 - Login',
